refactor: split MainViewModel into responsibility-scoped partials
The 4,370-line MainViewModel was a god class: connection, polling, register
grid editing, custom watch, project persistence, navigation, and window
commands all lived in one file. It is now seven partial files (same class,
zero behavior change, all 585 tests green):

- MainViewModel.cs (core) - lifecycle, DI fields, unit-configuration state
  and per-area config plumbing, child view-model properties, status
- MainViewModel.Connection.cs - connect/disconnect, auto-reconnect, profile
  events, unit-id parsing, connection status and failure counters
- MainViewModel.Polling.cs - the monitor poll loop, due-queue, area reads
- MainViewModel.Registers.cs - register/coil grid read/write, prompts,
  per-area metadata, legacy dual-state properties
- MainViewModel.CustomWatch.cs - custom entries, bulk add, watch loop
- MainViewModel.Project.cs - project save/load, unit-id import/export,
  workspace snapshots
- MainViewModel.Navigation.cs - tab visibility, navigation list, console
  and debug messages, theme
- MainViewModel.Windows.cs - menu/dialog commands

Decorative #region markers were dropped (they would have been orphaned).
The split was produced by a Roslyn member extractor so every member moved
verbatim (393 members in == 393 out, verified); the splitter script lives
in a scratch project and is not committed.
