+++ ModbusForge/ViewModels/MainViewModel.cs
@@ -21,84 +21,56 @@ using ModbusForge.Services;
 
 namespace ModbusForge.Avalonia.ViewModels
 {
-    public sealed partial class MainViewModel : ObservableObject, IDisposable
+    /// <summary>
+    /// MainViewModel - core members (lifecycle, unit configuration, status, child view models).
+    /// </summary>
+    public partial class MainViewModel : ObservableObject, IDisposable
+
     {
         private readonly IConnectionManager _connectionManager;
+
         private readonly ILogger<MainViewModel> _logger;
+
         private readonly IDispatcher _dispatcher;
+
         private readonly IUnitConfigurationStore _unitConfigurationStore;
+
         private readonly ICustomEntryService? _customEntryService;
+
         private readonly IFileDialogService? _fileDialogService;
+
         private readonly IFileSystem _fileSystem;
+
         private readonly IInputDialogService? _inputDialogService;
+
         private readonly ICustomBulkAddDialogService? _customBulkAddDialogService;
+
         private readonly IMessageBoxService? _messageBoxService;
+
         private readonly ISettingsService? _settingsService;
+
         private readonly IThemeService? _themeService;
+
         private readonly IUpdateService? _updateService;
+
         private readonly IWindowService? _windowService;
+
         private readonly IApplicationLifetime? _applicationLifetime;
+
         private readonly IDockingHost? _dockingHost;
+
         private readonly ITrendLogger? _trendLogger;
+
         private readonly MqttGatewayService? _mqttGateway;
+
         private readonly IConsoleLoggerService? _consoleLoggerService;
-        private CancellationTokenSource? _pollCts;
-        private readonly object _pollLifecycleLock = new();
-        private readonly object _pendingPollLock = new();
-        private readonly HashSet<PlcArea> _pendingPollAreas = new();
-        private CancellationTokenSource? _customWatchCts;
-        private readonly object _customWatchLifecycleLock = new();
+
         private readonly SemaphoreSlim _modbusIoGate = new(1, 1);
-        private readonly object _monitorFailureLock = new();
-        private readonly Dictionary<PlcArea, int> _monitorFailureCounts = new();
-        private readonly Dictionary<PlcArea, DateTime> _lastMonitorFailureUtc = new();
+
         private bool _disposed;
+
         private bool _isApplyingUnitConfiguration;
-        private CancellationTokenSource? _autoReconnectCts;
-        private readonly object _autoReconnectLock = new();
-        private volatile bool _userInitiatedDisconnect;
-        private DateTime _lastHoldingReadUtc;
-        private DateTime _lastInputRegReadUtc;
-        private DateTime _lastCoilsReadUtc;
-        private DateTime _lastDiscreteReadUtc;
 
-        private const int PollLoopIntervalMs = 50;
-        private const int MinimumMonitorPeriodMs = 50;
-        private const int DefaultCustomPeriodMs = 1000;
-        private const int MultiRegisterTypeIncrement = 2;
-        private const int SingleRegisterTypeIncrement = 1;
-
-        /// <summary>
-        /// Pause before restarting the poll loop after it died from an error, so a
-        /// persistently failing service cannot spin an error → immediate-restart loop.
-        /// </summary>
-        private const int PollRestartBackoffMs = 1000;
-
-        // TabItem indices in MainTabControl (MainView.axaml). Referenced by the navigation
-        // model and the open-tab commands instead of magic numbers.
-        private const int DashboardTabIndex = 0;
-        private const int TrendsTabIndex = 1;
-        private const int FrameInspectorTabIndex = 2;
-        private const int MqttTabIndex = 3;
-        private const int ScriptEditorTabIndex = 4;
-        private const int SignalGeneratorTabIndex = 5;
-        private const int SimulationTabIndex = 6;
-        private const int HoldingRegistersTabIndex = 7;
-        private const int CoilsTabIndex = 8;
-        private const int InputRegistersTabIndex = 9;
-        private const int DiscreteInputsTabIndex = 10;
-        private const int CustomWatchTabIndex = 11;
-        private const int DecodeTabIndex = 12;
-        private const int ConsoleTabIndex = 13;
-        private const int DebugTabIndex = 14;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadCommand))]
-        private int _startAddress = 0;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadCommand))]
-        private int _registerCount = 20;
 
         [ObservableProperty]
         [NotifyCanExecuteChangedFor(nameof(ConnectCommand))]
@@ -114,143 +86,13 @@ namespace ModbusForge.Avalonia.ViewModels
         [NotifyCanExecuteChangedFor(nameof(RemoveCustomEntryCommand))]
         private bool _isBusy;
 
+
         [ObservableProperty]
         private string _statusMessage = "Ready";
 
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadCommand))]
-        private bool _isContinuousRead = true;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadCommand))]
-        [NotifyCanExecuteChangedFor(nameof(WriteCommand))]
-        private PlcArea _selectedArea = PlcArea.HoldingRegister;
-
-        [ObservableProperty]
-        private int _selectedAreaIndex;
-
-        [ObservableProperty]
-        private string _globalType = "int";
-
-        [ObservableProperty]
-        private bool _swapBytes;
-
-        [ObservableProperty]
-        private bool _swapWords;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadHoldingRegistersCommand))]
-        private int _holdingRegisterStart = 1;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadHoldingRegistersCommand))]
-        private int _holdingRegisterCount = 20;
-
-        [ObservableProperty]
-        private string _registersGlobalType = "int";
-
-        [ObservableProperty]
-        private bool _registersSwapBytes;
-
-        [ObservableProperty]
-        private bool _registersSwapWords;
-
-        [ObservableProperty]
-        private bool _holdingMonitorEnabled = true;
-
-        [ObservableProperty]
-        private int _holdingMonitorPeriodMs = 1000;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadInputRegistersCommand))]
-        private int _inputRegisterStart = 1;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadInputRegistersCommand))]
-        private int _inputRegisterCount = 20;
-
-        [ObservableProperty]
-        private string _inputRegistersGlobalType = "int";
-
-        [ObservableProperty]
-        private bool _inputRegistersSwapBytes;
-
-        [ObservableProperty]
-        private bool _inputRegistersSwapWords;
-
-        [ObservableProperty]
-        private bool _inputRegistersMonitorEnabled;
-
-        [ObservableProperty]
-        private int _inputRegistersMonitorPeriodMs = 1000;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadCoilsCommand))]
-        private int _coilStart = 1;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadCoilsCommand))]
-        private int _coilCount = 20;
-
-        [ObservableProperty]
-        private bool _coilsMonitorEnabled;
-
-        [ObservableProperty]
-        private int _coilsMonitorPeriodMs = 1000;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadDiscreteInputsCommand))]
-        private int _discreteInputStart = 1;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadDiscreteInputsCommand))]
-        private int _discreteInputCount = 20;
-
-        [ObservableProperty]
-        private bool _discreteInputsMonitorEnabled;
-
-        [ObservableProperty]
-        private int _discreteInputsMonitorPeriodMs = 1000;
-
-        [ObservableProperty]
-        [NotifyPropertyChangedFor(nameof(UnitId))]
-        private ConnectionProfile? _activeProfile;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(AddCustomEntryCommand))]
-        [NotifyCanExecuteChangedFor(nameof(RemoveCustomEntryCommand))]
-        [NotifyCanExecuteChangedFor(nameof(SaveCustomCommand))]
-        [NotifyCanExecuteChangedFor(nameof(LoadCustomCommand))]
-        [NotifyCanExecuteChangedFor(nameof(SaveProjectCommand))]
-        [NotifyCanExecuteChangedFor(nameof(LoadProjectCommand))]
-        private bool _isCustomWatchMonitoring;
-
-        [ObservableProperty]
-        [NotifyCanExecuteChangedFor(nameof(ReadCustomEntryCommand))]
-        [NotifyCanExecuteChangedFor(nameof(WriteCustomEntryCommand))]
-        private CustomEntry? _selectedCustomEntry;
-
-        [ObservableProperty]
-        private ObservableCollection<RegisterEntry> _holdingRegisters = new();
-
-        [ObservableProperty]
-        private ObservableCollection<RegisterEntry> _inputRegisters = new();
-
-        [ObservableProperty]
-        private ObservableCollection<CoilEntry> _coils = new();
-
-        [ObservableProperty]
-        private ObservableCollection<CoilEntry> _discreteInputs = new();
-
-        public ObservableCollection<RegisterEntry> Registers => HoldingRegisters;
-
-        /// <summary>
-        /// Compatibility property for the custom-watch bindings. The collection now
-        /// belongs to the selected Unit ID configuration instead of one global list.
-        /// </summary>
-        public ObservableCollection<CustomEntry> CustomEntries => _unitConfigurationStore.CurrentConfig.CustomEntries;
 
         public IModbusService? ActiveService => _connectionManager.ActiveService;
+
 
         public int UnitId
         {
@@ -275,13 +117,15 @@ namespace ModbusForge.Avalonia.ViewModels
             }
         }
 
-        public bool IsRegisterArea => SelectedArea is PlcArea.HoldingRegister or PlcArea.InputRegister;
 
         public IReadOnlyList<string> RegisterTypes => RegisterEntry.AvailableTypes;
 
+
         public IReadOnlyList<string> CustomAreas => CustomEntry.AvailableAreas;
 
+
         public List<string> Modes { get; } = new() { "Client", "Server" };
+
 
         /// <summary>
         /// Index of the current <see cref="Mode"/> in <see cref="Modes"/>.
@@ -300,37 +144,25 @@ namespace ModbusForge.Avalonia.ViewModels
             }
         }
 
+
         /// <summary>
         /// The available server Unit IDs are owned by the shared store. The property is
         /// intentionally kept on the view model for existing bindings and callers.
         /// </summary>
         public ObservableCollection<byte> AvailableUnitIds => _unitConfigurationStore.AvailableUnitIds;
 
+
         /// <summary>
         /// The complete per-Unit ID state currently held by the shared store.
         /// </summary>
         public IReadOnlyDictionary<byte, UnitIdConfiguration> UnitConfigurations => _unitConfigurationStore.UnitConfigurations;
+
 
         /// <summary>
         /// The configuration selected in the shared Unit ID store.
         /// </summary>
         public UnitIdConfiguration CurrentConfig => _unitConfigurationStore.CurrentConfig;
 
-        public ObservableCollection<ConnectionProfile> ConnectionProfiles => _connectionManager.Profiles;
-
-        public ConnectionProfile? DashboardSelectedProfile
-        {
-            get => ActiveProfile;
-            set
-            {
-                if (value != null && !ReferenceEquals(_connectionManager.ActiveProfile, value))
-                {
-                    _connectionManager.SetActiveProfile(value);
-                }
-
-                OnPropertyChanged();
-            }
-        }
 
         public string Mode
         {
@@ -354,34 +186,9 @@ namespace ModbusForge.Avalonia.ViewModels
             }
         }
 
-        public bool IsServerMode => string.Equals(Mode, "Server", StringComparison.OrdinalIgnoreCase);
-
-        public bool ShowClientFields => !IsServerMode;
-
-        public bool ShowServerFields => IsServerMode;
-
-        public string ConnectButtonText => IsServerMode ? "Start Server" : "Connect";
-
-        public string ToggleConnectionButtonText => IsConnected ? "Disconnect" : ConnectButtonText;
-
-        public string ConnectionHeader => IsServerMode ? "Modbus Connection (Server)" : "Modbus Connection (Client)";
-
-        public string AddressLabel => IsServerMode ? "Interface:" : "Server:";
 
         public byte EffectiveUnitId => IsServerMode ? SelectedUnitId : (byte)UnitId;
 
-        public string ServerUnitIds
-        {
-            get => ActiveProfile?.ServerUnitIds ?? "1";
-            set
-            {
-                if (ActiveProfile != null && ActiveProfile.ServerUnitIds != value)
-                {
-                    ActiveProfile.ServerUnitIds = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
 
         public byte SelectedUnitId
         {
@@ -401,19 +208,27 @@ namespace ModbusForge.Avalonia.ViewModels
             }
         }
 
+
         public TrendViewModel? TrendViewModel { get; }
+
 
         public FrameInspectorViewModel? FrameInspectorViewModel { get; }
 
+
         public MqttViewModel? MqttViewModel { get; }
+
 
         public ScriptEditorViewModel? ScriptEditorViewModel { get; }
 
+
         public SignalGeneratorViewModel? SignalGeneratorViewModel { get; }
+
 
         public VisualNodeEditorViewModel? VisualNodeEditorViewModel { get; }
 
+
         public DecodeViewModel? DecodeViewModel { get; }
+
 
         public MainViewModel(
             IConnectionManager connectionManager,
@@ -617,224 +432,6 @@ namespace ModbusForge.Avalonia.ViewModels
             UpdateCustomWatchMonitoringState();
         }
 
-        public IAsyncRelayCommand ConnectCommand { get; }
-        public IAsyncRelayCommand DisconnectCommand { get; }
-        public IAsyncRelayCommand ToggleConnectionCommand { get; }
-        public IAsyncRelayCommand ReadCommand { get; }
-        public IAsyncRelayCommand WriteCommand { get; }
-        public IAsyncRelayCommand ReadHoldingRegistersCommand { get; }
-        public IAsyncRelayCommand ReadInputRegistersCommand { get; }
-        public IAsyncRelayCommand ReadCoilsCommand { get; }
-        public IAsyncRelayCommand ReadDiscreteInputsCommand { get; }
-        public IAsyncRelayCommand WriteHoldingRegisterCommand { get; }
-        public IAsyncRelayCommand WriteCoilCommand { get; }
-
-        public IAsyncRelayCommand AddCustomEntryCommand { get; }
-        public IAsyncRelayCommand AddBulkCustomEntryCommand { get; }
-        public IAsyncRelayCommand RemoveCustomEntryCommand { get; }
-        public IAsyncRelayCommand ReadCustomEntryCommand { get; }
-        public IAsyncRelayCommand WriteCustomEntryCommand { get; }
-        public IAsyncRelayCommand<CustomEntry?> ReadCustomNowCommand { get; }
-        public IAsyncRelayCommand<CustomEntry?> WriteCustomNowCommand { get; }
-        public IAsyncRelayCommand<CustomEntry?> DeleteCustomEntryCommand { get; }
-        public IAsyncRelayCommand ReadAllCustomNowCommand { get; }
-        public IAsyncRelayCommand SaveCustomCommand { get; }
-        public IAsyncRelayCommand LoadCustomCommand { get; }
-        public IAsyncRelayCommand SaveProjectCommand { get; }
-        public IAsyncRelayCommand LoadProjectCommand { get; }
-
-        public ICommand OpenPreferencesCommand { get; }
-        public ICommand OpenAboutCommand { get; }
-        public ICommand OpenHelpCommand { get; }
-        public ICommand OpenKeyboardShortcutsCommand { get; }
-        public ICommand OpenTroubleshootingCommand { get; }
-        public ICommand ToggleThemeCommand { get; }
-        public ICommand CheckForUpdatesCommand { get; }
-        public ICommand ExitCommand { get; }
-        public ICommand ReadShortcutCommand { get; }
-        public ICommand OpenTrendsCommand { get; }
-        public ICommand OpenFrameInspectorCommand { get; }
-        public ICommand OpenScriptEditorCommand { get; }
-        public ICommand OpenPcapCommand { get; }
-        public ICommand OpenTagBrowserCommand { get; }
-        public ICommand OpenWatchWindowCommand { get; }
-        public ICommand OpenConnectionManagerCommand { get; }
-        public IAsyncRelayCommand ImportUnitIdsCommand { get; }
-        public IAsyncRelayCommand ExportUnitIdsCommand { get; }
-        public IAsyncRelayCommand ImportUnitIdAsCommand { get; }
-        public IAsyncRelayCommand ExportUnitIdCommand { get; }
-        public IAsyncRelayCommand SaveAllConfigCommand { get; }
-        public ICommand ShowAllTabsCommand { get; }
-        public ICommand ResetTabsCommand { get; }
-        public ICommand ClearConsoleCommand { get; }
-        public ICommand ClearDebugCommand { get; }
-
-        public bool IsDarkMode
-        {
-            get => _themeService?.IsDarkMode ?? false;
-            set
-            {
-                if (_themeService != null && _themeService.IsDarkMode != value)
-                {
-                    _themeService.SetTheme(value);
-                }
-
-                OnPropertyChanged();
-            }
-        }
-
-        [ObservableProperty]
-        private int _selectedTabIndex;
-
-        /// <summary>
-        /// Left navigation list entries (see MainView.axaml). Only entries whose tab is
-        /// currently visible are present - hidden tabs are removed from the list, so the
-        /// view does not need per-item visibility bindings.
-        /// </summary>
-        public ObservableCollection<NavigationItem> NavigationItems { get; } = new();
-
-        private readonly List<NavigationItem> _allNavigationItems = new();
-
-        [ObservableProperty]
-        private NavigationItem? _selectedNavigationItem;
-
-        partial void OnSelectedNavigationItemChanged(NavigationItem? value)
-        {
-            if (value is not null && SelectedTabIndex != value.TabIndex)
-            {
-                SelectedTabIndex = value.TabIndex;
-            }
-        }
-
-        [ObservableProperty]
-        private bool _isRegistersTabVisible = true;
-
-        [ObservableProperty]
-        private bool _isInputRegistersTabVisible = true;
-
-        [ObservableProperty]
-        private bool _isCoilsTabVisible = true;
-
-        [ObservableProperty]
-        private bool _isDiscreteInputsTabVisible = true;
-
-        [ObservableProperty]
-        private bool _isCustomWatchTabVisible = true;
-
-        [ObservableProperty]
-        private bool _isSimulationTabVisible = true;
-
-        [ObservableProperty]
-        private bool _isDecodeTabVisible = true;
-
-        [ObservableProperty]
-        private bool _isTrendTabVisible = true;
-
-        [ObservableProperty]
-        private bool _isRegisterGridEditing;
-
-        [ObservableProperty]
-        private bool _isConsoleTabVisible = true;
-
-        [ObservableProperty]
-        private bool _isDebugTabVisible = true;
-
-        [ObservableProperty]
-        private bool _hasConnectionError;
-
-        /// <summary>
-        /// Console tab backing store. When the shared <see cref="IConsoleLoggerService"/> is
-        /// available (the normal DI case) this is the same collection the Modbus services,
-        /// script engine, and API server already log into - so the tab shows backend
-        /// messages too, and the configured MaxConsoleMessages cap applies.
-        /// </summary>
-        public ObservableCollection<string> ConsoleMessages { get; }
-
-        private readonly ObservableCollection<string> _consoleMessageFallback = new();
-
-        public ObservableCollection<string> DebugMessages { get; } = new();
-
-        public string VersionText => $"v{typeof(MainViewModel).Assembly.GetName().Version?.ToString(3) ?? "unknown"}";
-
-        public bool IsConnected => ActiveProfile?.IsConnected == true;
-
-        public bool IsDisconnected => !IsConnected && !HasConnectionError;
-
-        public bool IsConnectionErrorVisible => HasConnectionError && !IsConnected;
-
-        public string ServerIpAddress
-        {
-            get
-            {
-                if (!IsServerMode)
-                    return string.Empty;
-
-                var port = ActiveProfile?.Port ?? 502;
-                if (ActiveService is ModbusServerService server && IsConnected)
-                {
-                    return server.BoundEndpoint;
-                }
-
-                var ips = IpRangeHelper.GetAllLocalIPv4();
-                if (ips.Count == 0)
-                    return $"127.0.0.1:{port}";
-
-                var withPort = ips.Select(ip => $"{ip}:{port}");
-                return string.Join(", ", withPort);
-            }
-        }
-
-        public string ActiveProfileDisplayName
-        {
-            get
-            {
-                if (ActiveProfile == null)
-                    return "None";
-
-                if (IsServerMode)
-                    return $"{ActiveProfile.Name} ({ServerIpAddress})";
-
-                return ActiveProfile.DisplayName;
-            }
-        }
-
-        public DateTime LastErrorTime { get; private set; } = DateTime.MinValue;
-
-        public int HoldingMonitorFailureCount => GetMonitorFailureCount(PlcArea.HoldingRegister);
-
-        public int InputRegistersMonitorFailureCount => GetMonitorFailureCount(PlcArea.InputRegister);
-
-        public int CoilsMonitorFailureCount => GetMonitorFailureCount(PlcArea.Coil);
-
-        public int DiscreteInputsMonitorFailureCount => GetMonitorFailureCount(PlcArea.DiscreteInput);
-
-        public int GetMonitorFailureCount(PlcArea area)
-        {
-            lock (_monitorFailureLock)
-            {
-                return _monitorFailureCounts.TryGetValue(area, out var count) ? count : 0;
-            }
-        }
-
-        public DateTime GetLastMonitorFailureUtc(PlcArea area)
-        {
-            lock (_monitorFailureLock)
-            {
-                return _lastMonitorFailureUtc.TryGetValue(area, out var timestamp)
-                    ? timestamp
-                    : DateTime.MinValue;
-            }
-        }
-
-        public string ConnectionStatusText => IsConnected
-            ? (ActiveProfile?.Status ?? "Connected")
-            : HasConnectionError ? "Connection error" : "Not connected";
-
-        public string DebugSummary => $"Profile: {ActiveProfile?.DisplayName ?? "None"} | " +
-                                      $"Mode: {ActiveProfile?.Mode ?? "(null)"} | IsServer: {ActiveProfile?.IsServerMode} | " +
-                                      $"Connected: {IsConnected} | Busy: {IsBusy} | " +
-                                      $"Holding: {HoldingRegisters.Count} | Input: {InputRegisters.Count} | " +
-                                      $"Coils: {Coils.Count} | Discrete: {DiscreteInputs.Count}";
 
         partial void OnStatusMessageChanged(string value)
         {
@@ -842,6 +439,7 @@ namespace ModbusForge.Avalonia.ViewModels
             AppendDebugMessage($"{DateTime.Now:HH:mm:ss.fff} {value}");
             OnPropertyChanged(nameof(DebugSummary));
         }
+
 
         partial void OnHasConnectionErrorChanged(bool value)
         {
@@ -851,55 +449,6 @@ namespace ModbusForge.Avalonia.ViewModels
             OnPropertyChanged(nameof(DebugSummary));
         }
 
-        partial void OnIsRegistersTabVisibleChanged(bool value) => OnTabVisibilityChanged(HoldingRegistersTabIndex);
-        partial void OnIsInputRegistersTabVisibleChanged(bool value) => OnTabVisibilityChanged(InputRegistersTabIndex);
-        partial void OnIsCoilsTabVisibleChanged(bool value) => OnTabVisibilityChanged(CoilsTabIndex);
-        partial void OnIsDiscreteInputsTabVisibleChanged(bool value) => OnTabVisibilityChanged(DiscreteInputsTabIndex);
-        partial void OnIsCustomWatchTabVisibleChanged(bool value) => OnTabVisibilityChanged(CustomWatchTabIndex);
-        partial void OnIsSimulationTabVisibleChanged(bool value) => OnTabVisibilityChanged(SimulationTabIndex);
-        partial void OnIsDecodeTabVisibleChanged(bool value) => OnTabVisibilityChanged(DecodeTabIndex);
-        partial void OnIsTrendTabVisibleChanged(bool value) => OnTabVisibilityChanged(TrendsTabIndex);
-        partial void OnIsConsoleTabVisibleChanged(bool value) => OnTabVisibilityChanged(ConsoleTabIndex);
-        partial void OnIsDebugTabVisibleChanged(bool value) => OnTabVisibilityChanged(DebugTabIndex);
-
-        private string? _lastConsoleMessage;
-
-        private void AppendConsoleMessage(string message)
-        {
-            if (string.IsNullOrWhiteSpace(message)) return;
-
-            // The poll loop re-sets the same status message every cycle; appending every
-            // repetition is what used to spam the console panel.
-            if (message == _lastConsoleMessage) return;
-            _lastConsoleMessage = message;
-
-            if (_consoleLoggerService != null)
-            {
-                // Shared sink: dispatches to the UI thread and enforces the configured
-                // MaxConsoleMessages cap.
-                _consoleLoggerService.Log(message);
-                return;
-            }
-
-            // Fallback (no DI-provided console service, e.g. some test setups).
-            var cap = Math.Max(1, _settingsService?.MaxConsoleMessages ?? 1000);
-            ConsoleMessages.Add(message);
-            while (ConsoleMessages.Count > cap)
-            {
-                ConsoleMessages.RemoveAt(0);
-            }
-        }
-
-        private void AppendDebugMessage(string message)
-        {
-            if (string.IsNullOrWhiteSpace(message)) return;
-
-            DebugMessages.Add(message);
-            while (DebugMessages.Count > 1000)
-            {
-                DebugMessages.RemoveAt(0);
-            }
-        }
 
         private void OnVisualNodeEditorViewModelPropertyChanged(object? sender, PropertyChangedEventArgs e)
         {
@@ -915,243 +464,6 @@ namespace ModbusForge.Avalonia.ViewModels
             }
         }
 
-        public void ShowAllTabs()
-        {
-            IsRegistersTabVisible = true;
-            IsInputRegistersTabVisible = true;
-            IsCoilsTabVisible = true;
-            IsDiscreteInputsTabVisible = true;
-            IsCustomWatchTabVisible = true;
-            IsSimulationTabVisible = true;
-            IsDecodeTabVisible = true;
-            IsTrendTabVisible = true;
-            IsConsoleTabVisible = true;
-            IsDebugTabVisible = true;
-        }
-
-        public void ResetTabs() => ShowAllTabs();
-
-        public List<string> GetVisibleTabs()
-        {
-            var visibleTabs = new List<string>();
-            if (IsRegistersTabVisible) visibleTabs.Add("Registers");
-            if (IsInputRegistersTabVisible) visibleTabs.Add("InputRegisters");
-            if (IsCoilsTabVisible) visibleTabs.Add("Coils");
-            if (IsDiscreteInputsTabVisible) visibleTabs.Add("DiscreteInputs");
-            if (IsCustomWatchTabVisible) visibleTabs.Add("CustomWatch");
-            if (IsSimulationTabVisible) visibleTabs.Add("Simulation");
-            if (IsDecodeTabVisible) visibleTabs.Add("Decode");
-            if (IsTrendTabVisible) visibleTabs.Add("Trend");
-            if (IsConsoleTabVisible) visibleTabs.Add("Console");
-            if (IsDebugTabVisible) visibleTabs.Add("Debug");
-            return visibleTabs;
-        }
-
-        public void SetVisibleTabs(IReadOnlyCollection<string>? visibleTabs)
-        {
-            if (visibleTabs == null || visibleTabs.Count == 0)
-            {
-                ShowAllTabs();
-                return;
-            }
-
-            IsRegistersTabVisible = visibleTabs.Contains("Registers");
-            IsInputRegistersTabVisible = visibleTabs.Contains("InputRegisters");
-            IsCoilsTabVisible = visibleTabs.Contains("Coils");
-            IsDiscreteInputsTabVisible = visibleTabs.Contains("DiscreteInputs");
-            IsCustomWatchTabVisible = visibleTabs.Contains("CustomWatch");
-            IsSimulationTabVisible = visibleTabs.Contains("Simulation");
-            IsDecodeTabVisible = visibleTabs.Contains("Decode");
-            IsTrendTabVisible = visibleTabs.Contains("Trend");
-            IsConsoleTabVisible = visibleTabs.Contains("Console");
-            IsDebugTabVisible = visibleTabs.Contains("Debug");
-        }
-
-        private void EnsureSelectedTabIsVisible()
-        {
-            if (IsTabIndexVisible(SelectedTabIndex)) return;
-
-            SelectedTabIndex = Enumerable.Range(0, _allNavigationItems.Count).FirstOrDefault(IsTabIndexVisible);
-        }
-
-        private bool IsTabIndexVisible(int index)
-        {
-            return index switch
-            {
-                TrendsTabIndex => IsTrendTabVisible,
-                SimulationTabIndex => IsSimulationTabVisible,
-                HoldingRegistersTabIndex => IsRegistersTabVisible,
-                CoilsTabIndex => IsCoilsTabVisible,
-                InputRegistersTabIndex => IsInputRegistersTabVisible,
-                DiscreteInputsTabIndex => IsDiscreteInputsTabVisible,
-                CustomWatchTabIndex => IsCustomWatchTabVisible,
-                DecodeTabIndex => IsDecodeTabVisible,
-                ConsoleTabIndex => IsConsoleTabVisible,
-                DebugTabIndex => IsDebugTabVisible,
-                _ => true
-            };
-        }
-
-        private void ThemeService_ThemeChanged(object? sender, EventArgs e)
-        {
-            OnPropertyChanged(nameof(IsDarkMode));
-        }
-
-        partial void OnSelectedAreaChanged(PlcArea value)
-        {
-            IsRegisterGridEditing = false;
-            SelectedAreaIndex = (int)value;
-            OnPropertyChanged(nameof(IsRegisterArea));
-            OnPropertyChanged(nameof(CanWrite));
-            WriteCommand.NotifyCanExecuteChanged();
-
-            // Keep legacy global properties in sync with the newly selected area
-            StartAddress = GetAreaStart(value);
-            RegisterCount = GetAreaCount(value);
-            GlobalType = GetAreaGlobalType(value);
-            SwapBytes = GetAreaSwapBytes(value);
-            SwapWords = GetAreaSwapWords(value);
-        }
-
-        partial void OnSelectedTabIndexChanged(int value)
-        {
-            IsRegisterGridEditing = false;
-
-            // Keep the navigation list in sync when the tab is changed programmatically.
-            var item = _allNavigationItems.FirstOrDefault(i => i.TabIndex == value && i.IsVisible);
-            if (!ReferenceEquals(item, SelectedNavigationItem))
-            {
-                SelectedNavigationItem = item;
-            }
-        }
-
-        /// <summary>
-        /// Shared handler for the tab-visibility flags: re-anchors the selection if the
-        /// current tab was hidden and updates the (filtered) navigation list.
-        /// </summary>
-        private void OnTabVisibilityChanged(int tabIndex)
-        {
-            EnsureSelectedTabIsVisible();
-            RefreshNavigationItems();
-
-            if (SelectedNavigationItem is { IsVisible: false })
-            {
-                // The selected entry just disappeared from the list - re-select to match
-                // the (re-anchored) tab.
-                OnSelectedTabIndexChanged(SelectedTabIndex);
-            }
-        }
-
-        /// <summary>Rebuilds the visible subset of the navigation list in place.</summary>
-        private void RefreshNavigationItems()
-        {
-            foreach (var item in _allNavigationItems)
-            {
-                if (item.IsVisible && !NavigationItems.Contains(item))
-                {
-                    NavigationItems.Add(item);
-                }
-                else if (!item.IsVisible)
-                {
-                    NavigationItems.Remove(item);
-                }
-            }
-        }
-
-        partial void OnSelectedAreaIndexChanged(int value)
-        {
-            SelectedArea = (PlcArea)value;
-        }
-
-        partial void OnIsContinuousReadChanged(bool value)
-        {
-            HoldingMonitorEnabled = value;
-        }
-
-        partial void OnStartAddressChanged(int value)
-        {
-            SetAreaStart(SelectedArea, value);
-        }
-
-        partial void OnRegisterCountChanged(int value)
-        {
-            SetAreaCount(SelectedArea, value);
-        }
-
-        partial void OnGlobalTypeChanged(string value)
-        {
-            SetAreaGlobalType(SelectedArea, value);
-        }
-
-        partial void OnSwapBytesChanged(bool value)
-        {
-            SetAreaSwapBytes(SelectedArea, value);
-        }
-
-        partial void OnSwapWordsChanged(bool value)
-        {
-            SetAreaSwapWords(SelectedArea, value);
-        }
-
-        partial void OnHoldingMonitorEnabledChanged(bool value)
-        {
-            if (value)
-            {
-                StartPolling();
-            }
-            else if (!AnyMonitorEnabled())
-            {
-                StopPolling();
-            }
-        }
-
-        partial void OnInputRegistersMonitorEnabledChanged(bool value)
-        {
-            if (value)
-            {
-                StartPolling();
-            }
-            else if (!AnyMonitorEnabled())
-            {
-                StopPolling();
-            }
-        }
-
-        partial void OnCoilsMonitorEnabledChanged(bool value)
-        {
-            if (value)
-            {
-                StartPolling();
-            }
-            else if (!AnyMonitorEnabled())
-            {
-                StopPolling();
-            }
-        }
-
-        partial void OnDiscreteInputsMonitorEnabledChanged(bool value)
-        {
-            if (value)
-            {
-                StartPolling();
-            }
-            else if (!AnyMonitorEnabled())
-            {
-                StopPolling();
-            }
-        }
-
-        partial void OnIsCustomWatchMonitoringChanged(bool value)
-        {
-            if (value)
-            {
-                StartCustomWatchMonitoring();
-            }
-            else
-            {
-                StopCustomWatchMonitoring();
-            }
-        }
 
         protected override void OnPropertyChanged(PropertyChangedEventArgs e)
         {
@@ -1164,84 +476,6 @@ namespace ModbusForge.Avalonia.ViewModels
             }
         }
 
-        private ObservableCollection<CustomEntry>? _hookedCustomEntries;
-
-        private void HookCustomEntries()
-        {
-            if (_hookedCustomEntries == CustomEntries) return;
-
-            UnhookCustomEntries();
-
-            _hookedCustomEntries = CustomEntries;
-
-            if (_hookedCustomEntries != null)
-            {
-                _hookedCustomEntries.CollectionChanged += OnCustomEntriesCollectionChanged;
-                foreach (var entry in _hookedCustomEntries)
-                {
-                    entry.PropertyChanged += OnCustomEntryPropertyChanged;
-                }
-            }
-        }
-
-        private void UnhookCustomEntries()
-        {
-            if (_hookedCustomEntries == null)
-                return;
-
-            _hookedCustomEntries.CollectionChanged -= OnCustomEntriesCollectionChanged;
-            foreach (var entry in _hookedCustomEntries)
-            {
-                entry.PropertyChanged -= OnCustomEntryPropertyChanged;
-            }
-            _hookedCustomEntries = null;
-        }
-
-        private void OnCustomEntriesCollectionChanged(object? sender, NotifyCollectionChangedEventArgs e)
-        {
-            if (e.NewItems != null)
-            {
-                foreach (CustomEntry item in e.NewItems)
-                {
-                    item.PropertyChanged += OnCustomEntryPropertyChanged;
-                }
-            }
-
-            if (e.OldItems != null)
-            {
-                foreach (CustomEntry item in e.OldItems)
-                {
-                    item.PropertyChanged -= OnCustomEntryPropertyChanged;
-                }
-            }
-
-            UpdateCustomWatchMonitoringState();
-        }
-
-        private void OnCustomEntryPropertyChanged(object? sender, PropertyChangedEventArgs e)
-        {
-            if (e.PropertyName is nameof(CustomEntry.Monitor) or nameof(CustomEntry.Continuous))
-            {
-                UpdateCustomWatchMonitoringState();
-            }
-        }
-
-        private void UpdateCustomWatchMonitoringState()
-        {
-            var shouldMonitor = CustomEntries.Any(entry => entry.Monitor || entry.Continuous);
-            if (IsCustomWatchMonitoring != shouldMonitor)
-            {
-                IsCustomWatchMonitoring = shouldMonitor;
-            }
-        }
-
-        partial void OnActiveProfileChanged(ConnectionProfile? value)
-        {
-            RefreshConnectionDependentProperties();
-            ConnectCommand.NotifyCanExecuteChanged();
-            DisconnectCommand.NotifyCanExecuteChanged();
-            ToggleConnectionCommand.NotifyCanExecuteChanged();
-        }
 
         partial void OnIsBusyChanged(bool value)
         {
@@ -1262,62 +496,9 @@ namespace ModbusForge.Avalonia.ViewModels
             ImportUnitIdAsCommand.NotifyCanExecuteChanged();
         }
 
-        private void RefreshConnectionDependentProperties()
-        {
-            if (ActiveProfile is null) return;
-
-            RefreshAvailableUnitIds();
-            SyncCurrentUnitConfiguration();
-
-            OnPropertyChanged(nameof(ActiveService));
-            OnPropertyChanged(nameof(ServerIpAddress));
-            OnPropertyChanged(nameof(ActiveProfileDisplayName));
-            OnPropertyChanged(nameof(UnitId));
-            OnPropertyChanged(nameof(Mode));
-            OnPropertyChanged(nameof(ModeIndex));
-            OnPropertyChanged(nameof(IsServerMode));
-            OnPropertyChanged(nameof(ShowClientFields));
-            OnPropertyChanged(nameof(ShowServerFields));
-            OnPropertyChanged(nameof(ConnectButtonText));
-            OnPropertyChanged(nameof(ToggleConnectionButtonText));
-            OnPropertyChanged(nameof(ConnectionHeader));
-            OnPropertyChanged(nameof(AddressLabel));
-            OnPropertyChanged(nameof(ServerUnitIds));
-            OnPropertyChanged(nameof(AvailableUnitIds));
-            OnPropertyChanged(nameof(SelectedUnitId));
-            OnPropertyChanged(nameof(EffectiveUnitId));
-            OnPropertyChanged(nameof(CanConnect));
-            OnPropertyChanged(nameof(CanDisconnect));
-            OnPropertyChanged(nameof(CanRead));
-            OnPropertyChanged(nameof(CanWrite));
-            OnPropertyChanged(nameof(CanReadCustomEntry));
-            OnPropertyChanged(nameof(CanWriteCustomEntry));
-        }
-
-        private bool CanConnect() => ActiveProfile is { IsConnected: false } && !IsBusy;
-
-        private bool CanDisconnect() => ActiveProfile is { IsConnected: true } && !IsBusy;
-
-        private bool CanRead() => CanRead(SelectedArea);
-
-        private bool CanRead(PlcArea area)
-        {
-            if (ActiveProfile is not { IsConnected: true } || IsBusy)
-                return false;
-
-            var (start, count) = GetAreaStartCount(area);
-            var validator = new ModbusAddressValidator();
-            return area is PlcArea.HoldingRegister or PlcArea.InputRegister
-                ? validator.IsValidAddressRange(start, count)
-                : validator.IsValidRange(start, count, area);
-        }
-
-        private bool CanWrite() => CanWrite(SelectedArea);
-
-        private bool CanWrite(PlcArea area) => ActiveProfile is { IsConnected: true } && !IsBusy &&
-                                               (area is PlcArea.HoldingRegister or PlcArea.Coil);
 
         private (int Start, int Count) GetAreaStartCount(PlcArea area) => (GetAreaStart(area), GetAreaCount(area));
+
 
         private int GetAreaStart(PlcArea area)
         {
@@ -1331,6 +512,7 @@ namespace ModbusForge.Avalonia.ViewModels
             };
         }
 
+
         private int GetAreaCount(PlcArea area)
         {
             return area switch
@@ -1343,6 +525,7 @@ namespace ModbusForge.Avalonia.ViewModels
             };
         }
 
+
         private string GetAreaGlobalType(PlcArea area)
         {
             return area switch
@@ -1352,6 +535,7 @@ namespace ModbusForge.Avalonia.ViewModels
                 _ => "int"
             };
         }
+
 
         private bool GetAreaSwapBytes(PlcArea area)
         {
@@ -1363,6 +547,7 @@ namespace ModbusForge.Avalonia.ViewModels
             };
         }
 
+
         private bool GetAreaSwapWords(PlcArea area)
         {
             return area switch
@@ -1372,6 +557,7 @@ namespace ModbusForge.Avalonia.ViewModels
                 _ => false
             };
         }
+
 
         private void SetAreaStart(PlcArea area, int value)
         {
@@ -1384,6 +570,7 @@ namespace ModbusForge.Avalonia.ViewModels
             }
         }
 
+
         private void SetAreaCount(PlcArea area, int value)
         {
             switch (area)
@@ -1395,11 +582,13 @@ namespace ModbusForge.Avalonia.ViewModels
             }
         }
 
+
         private void SetAreaGlobalType(PlcArea area, string value)
         {
             if (area == PlcArea.HoldingRegister) RegistersGlobalType = value;
             else if (area == PlcArea.InputRegister) InputRegistersGlobalType = value;
         }
+
 
         private void SetAreaSwapBytes(PlcArea area, bool value)
         {
@@ -1407,37 +596,13 @@ namespace ModbusForge.Avalonia.ViewModels
             else if (area == PlcArea.InputRegister) InputRegistersSwapBytes = value;
         }
 
+
         private void SetAreaSwapWords(PlcArea area, bool value)
         {
             if (area == PlcArea.HoldingRegister) RegistersSwapWords = value;
             else if (area == PlcArea.InputRegister) InputRegistersSwapWords = value;
         }
 
-        private bool AnyMonitorEnabled() => HoldingMonitorEnabled || InputRegistersMonitorEnabled || CoilsMonitorEnabled || DiscreteInputsMonitorEnabled;
-
-        private bool CanRemoveCustomEntry() => SelectedCustomEntry != null && !IsBusy;
-
-        private bool CanReadCustomEntry() => SelectedCustomEntry != null && ActiveProfile is { IsConnected: true } && !IsBusy;
-
-        private bool CanReadCustomEntry(CustomEntry? entry) => entry != null && ActiveProfile is { IsConnected: true } && !IsBusy;
-
-        private bool CanWriteCustomEntry() => SelectedCustomEntry != null && ActiveProfile is { IsConnected: true } && !IsBusy;
-
-        private bool CanWriteCustomEntry(CustomEntry? entry) => entry != null && ActiveProfile is { IsConnected: true } && !IsBusy;
-
-        private bool CanReadAllCustomEntries() => CustomEntries.Count > 0 && ActiveProfile is { IsConnected: true } && !IsBusy;
-
-        private bool CanSaveCustom() => _customEntryService != null && !IsBusy;
-
-        private bool CanLoadCustom() => _customEntryService != null && !IsBusy;
-
-        private bool CanSaveProject() => _fileDialogService != null && !IsBusy;
-
-        private bool CanLoadProject() => _fileDialogService != null && !IsBusy;
-
-        private bool CanExportUnitId() => _fileDialogService != null && IsServerMode && !IsBusy;
-
-        private bool CanImportUnitIdAs() => _fileDialogService != null && _inputDialogService != null && IsServerMode && !IsBusy;
 
         private void UnitConfigurationStore_SelectedUnitIdChanged(object? sender, EventArgs e)
         {
@@ -1458,12 +623,14 @@ namespace ModbusForge.Avalonia.ViewModels
             ImportUnitIdAsCommand.NotifyCanExecuteChanged();
         }
 
+
         private void UnitConfigurationStore_AvailableUnitIdsChanged(object? sender, EventArgs e)
         {
             OnPropertyChanged(nameof(AvailableUnitIds));
             OnPropertyChanged(nameof(UnitConfigurations));
             OnPropertyChanged(nameof(SelectedUnitId));
         }
+
 
         /// <summary>
         /// Copies the live Avalonia controls' state into the selected Unit ID
@@ -1541,6 +708,7 @@ namespace ModbusForge.Avalonia.ViewModels
             }
         }
 
+
         /// <summary>
         /// Applies the selected Unit ID's persisted settings to the existing
         /// Avalonia controls without replacing the live collections.
@@ -1592,6 +760,7 @@ namespace ModbusForge.Avalonia.ViewModels
             ReadAllCustomNowCommand.NotifyCanExecuteChanged();
         }
 
+
         private static void ApplyRegisterMetadata(
             IEnumerable<RegisterEntry> entries,
             IEnumerable<RegisterMetadata>? metadata)
@@ -1615,6 +784,7 @@ namespace ModbusForge.Avalonia.ViewModels
             }
         }
 
+
         private void ActiveProfile_PropertyChanged(object? sender, PropertyChangedEventArgs e)
         {
             // The profile is mutated from thread-pool threads too (ConnectProfileAsync /
@@ -1623,1917 +793,6 @@ namespace ModbusForge.Avalonia.ViewModels
             _ = _dispatcher.InvokeAsync(() => HandleActiveProfilePropertyChanged(e));
         }
 
-        private void HandleActiveProfilePropertyChanged(PropertyChangedEventArgs e)
-        {
-            if (e.PropertyName == nameof(ConnectionProfile.IsConnected))
-            {
-                if (ActiveProfile?.IsConnected == true)
-                {
-                    HasConnectionError = false;
-                }
-
-                OnPropertyChanged(nameof(IsConnected));
-                OnPropertyChanged(nameof(IsDisconnected));
-                OnPropertyChanged(nameof(IsConnectionErrorVisible));
-                OnPropertyChanged(nameof(ConnectionStatusText));
-                OnPropertyChanged(nameof(ServerIpAddress));
-            OnPropertyChanged(nameof(ActiveProfileDisplayName));
-                OnPropertyChanged(nameof(DebugSummary));
-                OnPropertyChanged(nameof(CanConnect));
-                OnPropertyChanged(nameof(CanDisconnect));
-                OnPropertyChanged(nameof(ToggleConnectionButtonText));
-                OnPropertyChanged(nameof(CanRead));
-                OnPropertyChanged(nameof(CanWrite));
-                OnPropertyChanged(nameof(CanReadCustomEntry));
-                OnPropertyChanged(nameof(CanWriteCustomEntry));
-                ConnectCommand.NotifyCanExecuteChanged();
-                ToggleConnectionCommand.NotifyCanExecuteChanged();
-                DisconnectCommand.NotifyCanExecuteChanged();
-                ReadCommand.NotifyCanExecuteChanged();
-                WriteCommand.NotifyCanExecuteChanged();
-                ReadHoldingRegistersCommand.NotifyCanExecuteChanged();
-                ReadInputRegistersCommand.NotifyCanExecuteChanged();
-                ReadCoilsCommand.NotifyCanExecuteChanged();
-                ReadDiscreteInputsCommand.NotifyCanExecuteChanged();
-                WriteHoldingRegisterCommand.NotifyCanExecuteChanged();
-                WriteCoilCommand.NotifyCanExecuteChanged();
-                ReadCustomEntryCommand.NotifyCanExecuteChanged();
-                WriteCustomEntryCommand.NotifyCanExecuteChanged();
-                ReadCustomNowCommand.NotifyCanExecuteChanged();
-                WriteCustomNowCommand.NotifyCanExecuteChanged();
-                ReadAllCustomNowCommand.NotifyCanExecuteChanged();
-            }
-
-            if (e.PropertyName == nameof(ConnectionProfile.Status))
-            {
-                StatusMessage = ActiveProfile?.Status ?? "Ready";
-                OnPropertyChanged(nameof(ConnectionStatusText));
-            }
-
-            if (e.PropertyName == nameof(ConnectionProfile.Port))
-            {
-                OnPropertyChanged(nameof(ServerIpAddress));
-            OnPropertyChanged(nameof(ActiveProfileDisplayName));
-            }
-
-            if (e.PropertyName == nameof(ConnectionProfile.UnitId))
-            {
-                if (!IsServerMode && ActiveProfile != null)
-                {
-                    _unitConfigurationStore.GetOrCreateConfiguration(ActiveProfile.UnitId);
-                    if (_unitConfigurationStore.SelectedUnitId != ActiveProfile.UnitId)
-                    {
-                        _unitConfigurationStore.SelectedUnitId = ActiveProfile.UnitId;
-                    }
-                }
-
-                OnPropertyChanged(nameof(UnitId));
-                OnPropertyChanged(nameof(EffectiveUnitId));
-            }
-
-            if (e.PropertyName == nameof(ConnectionProfile.Mode))
-            {
-                OnPropertyChanged(nameof(Mode));
-                OnPropertyChanged(nameof(IsServerMode));
-                OnPropertyChanged(nameof(ShowClientFields));
-                OnPropertyChanged(nameof(ShowServerFields));
-                OnPropertyChanged(nameof(ConnectButtonText));
-                OnPropertyChanged(nameof(ToggleConnectionButtonText));
-                OnPropertyChanged(nameof(ConnectionHeader));
-                OnPropertyChanged(nameof(AddressLabel));
-                OnPropertyChanged(nameof(EffectiveUnitId));
-                OnPropertyChanged(nameof(DebugSummary));
-                RefreshAvailableUnitIds();
-                OnPropertyChanged(nameof(AvailableUnitIds));
-                OnPropertyChanged(nameof(SelectedUnitId));
-                ExportUnitIdCommand.NotifyCanExecuteChanged();
-                ImportUnitIdAsCommand.NotifyCanExecuteChanged();
-            }
-
-            if (e.PropertyName == nameof(ConnectionProfile.ServerUnitIds))
-            {
-                OnPropertyChanged(nameof(ServerUnitIds));
-                RefreshAvailableUnitIds();
-            }
-        }
-
-        private void ConnectionManager_ActiveProfileChanged(object? sender, ConnectionProfile? e)
-        {
-            // Raised on the caller's thread (UI in most cases, but ConnectionManager can be
-            // driven from other threads) - marshal to keep VM state on the UI thread.
-            _ = _dispatcher.InvokeAsync(() => HandleActiveProfileChanged(e));
-        }
-
-        private void HandleActiveProfileChanged(ConnectionProfile? e)
-        {
-            // The auto-reconnect loop targets a specific profile; a profile switch voids it.
-            StopAutoReconnect();
-
-            SyncCurrentUnitConfiguration();
-
-            if (ActiveProfile != null)
-            {
-                ActiveProfile.PropertyChanged -= ActiveProfile_PropertyChanged;
-            }
-
-            ActiveProfile = e;
-
-            if (e != null)
-            {
-                e.PropertyChanged += ActiveProfile_PropertyChanged;
-                if (!e.IsServerMode)
-                {
-                    SelectedUnitId = e.UnitId;
-                }
-            }
-
-            OnPropertyChanged(nameof(ActiveProfile));
-            OnPropertyChanged(nameof(DashboardSelectedProfile));
-            RefreshConnectionDependentProperties();
-            ReadCommand.NotifyCanExecuteChanged();
-            WriteCommand.NotifyCanExecuteChanged();
-            ReadCustomEntryCommand.NotifyCanExecuteChanged();
-            WriteCustomEntryCommand.NotifyCanExecuteChanged();
-            ReadCustomNowCommand.NotifyCanExecuteChanged();
-            WriteCustomNowCommand.NotifyCanExecuteChanged();
-            ReadAllCustomNowCommand.NotifyCanExecuteChanged();
-
-            RefreshAvailableUnitIds();
-
-            ResetMonitorFailures();
-            HasConnectionError = false;
-            OnPropertyChanged(nameof(IsConnected));
-            OnPropertyChanged(nameof(IsDisconnected));
-            OnPropertyChanged(nameof(IsConnectionErrorVisible));
-            OnPropertyChanged(nameof(ConnectionStatusText));
-            OnPropertyChanged(nameof(DebugSummary));
-            StatusMessage = e != null ? $"Active profile: {e.DisplayName}" : "No active connection profile";
-        }
-
-        private void ConnectionManager_ProfileConnected(object? sender, ConnectionProfile e)
-        {
-            // Raised from ConnectProfileAsync on a thread-pool thread.
-            _ = _dispatcher.InvokeAsync(() => HandleProfileConnected(e));
-        }
-
-        private void HandleProfileConnected(ConnectionProfile e)
-        {
-            // A successful connection (manual or auto) ends any pending auto-reconnect.
-            StopAutoReconnect();
-
-            _logger.LogInformation("Profile connected: {Name}", e.Name);
-            ResetMonitorFailures();
-            HasConnectionError = false;
-            OnPropertyChanged(nameof(IsConnected));
-            OnPropertyChanged(nameof(IsDisconnected));
-            OnPropertyChanged(nameof(IsConnectionErrorVisible));
-            OnPropertyChanged(nameof(ConnectionStatusText));
-            OnPropertyChanged(nameof(ServerIpAddress));
-            OnPropertyChanged(nameof(ActiveProfileDisplayName));
-            OnPropertyChanged(nameof(DebugSummary));
-
-            if (e.IsServerMode && _connectionManager.ActiveService is ModbusServerService server)
-            {
-                var unitIds = server.GetUnitIds()
-                    .Where(id => id is >= 1 and <= 247)
-                    .Distinct()
-                    .OrderBy(id => id)
-                    .ToList();
-                _unitConfigurationStore.PopulateAvailableUnitIds(unitIds);
-
-                foreach (var id in unitIds)
-                {
-                    _unitConfigurationStore.GetOrCreateConfiguration(id);
-                }
-
-                if (unitIds.Count > 0)
-                {
-                    SelectedUnitId = unitIds[0];
-                }
-
-                StatusMessage = $"Server started on {server.BoundEndpoint}";
-
-                OnPropertyChanged(nameof(ShowServerFields));
-                ExportUnitIdCommand.NotifyCanExecuteChanged();
-                ImportUnitIdAsCommand.NotifyCanExecuteChanged();
-            }
-
-            _trendLogger?.Start();
-            StartPolling();
-            StartCustomWatchMonitoring();
-        }
-
-        private void ConnectionManager_ProfileDisconnected(object? sender, ConnectionProfile e)
-        {
-            // Raised from DisconnectProfileAsync on a thread-pool thread.
-            _ = _dispatcher.InvokeAsync(() => HandleProfileDisconnected(e));
-        }
-
-        private void HandleProfileDisconnected(ConnectionProfile e)
-        {
-            _logger.LogInformation("Profile disconnected: {Name}", e.Name);
-            OnPropertyChanged(nameof(ServerIpAddress));
-            OnPropertyChanged(nameof(ActiveProfileDisplayName));
-            _trendLogger?.Stop();
-            StopPolling();
-            StopCustomWatchMonitoring();
-
-            var userInitiated = _userInitiatedDisconnect;
-            _userInitiatedDisconnect = false;
-
-            // Profiles removed from the manager (profile deleted, project reloaded) must
-            // never be reconnected either.
-            if (userInitiated || !_connectionManager.Profiles.Contains(e))
-            {
-                StopAutoReconnect();
-            }
-            else
-            {
-                // Unexpected loss: restart the link if the user opted in.
-                StartAutoReconnectIfEnabled("unexpected connection loss");
-            }
-        }
-
-        private void RefreshAvailableUnitIds()
-        {
-            if (ActiveProfile is null || !IsServerMode)
-            {
-                _unitConfigurationStore.PopulateAvailableUnitIds(Array.Empty<byte>());
-                return;
-            }
-
-            var ids = ParseUnitIdString(ActiveProfile.ServerUnitIds);
-            _unitConfigurationStore.PopulateAvailableUnitIds(ids);
-
-            foreach (var id in ids)
-            {
-                _unitConfigurationStore.GetOrCreateConfiguration(id);
-            }
-
-            if (ids.Count > 0 && (SelectedUnitId == 0 || !ids.Contains(SelectedUnitId)))
-            {
-                SelectedUnitId = ids[0];
-            }
-        }
-
-        private static List<byte> ParseUnitIdString(string? input)
-        {
-            var result = new List<byte>();
-            if (string.IsNullOrWhiteSpace(input)) return result;
-
-            var parts = input.Split(new[] { ',', ';' }, StringSplitOptions.RemoveEmptyEntries);
-            foreach (var part in parts)
-            {
-                var trimmed = part.Trim();
-                if (trimmed.Contains('-'))
-                {
-                    var range = trimmed.Split('-');
-                    if (range.Length == 2 && byte.TryParse(range[0].Trim(), out byte start) && byte.TryParse(range[1].Trim(), out byte end))
-                    {
-                        for (int i = Math.Min(start, end); i <= Math.Max(start, end); i++)
-                        {
-                            if (i >= 1 && i <= 247 && !result.Contains((byte)i))
-                                result.Add((byte)i);
-                        }
-                    }
-                }
-                else if (byte.TryParse(trimmed, out byte id))
-                {
-                    if (id >= 1 && id <= 247 && !result.Contains(id))
-                        result.Add(id);
-                }
-            }
-
-            return result;
-        }
-
-        private async Task ConnectAsync()
-        {
-            if (ActiveProfile == null) return;
-
-            IsBusy = true;
-            try
-            {
-                await _connectionManager.ConnectProfileAsync(ActiveProfile);
-            }
-            catch (Exception ex)
-            {
-                HasConnectionError = true;
-                StatusMessage = $"Connection error: {ex.Message}";
-                _logger.LogError(ex, "Error connecting profile {Name}", ActiveProfile.Name);
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        private async Task DisconnectAsync()
-        {
-            if (ActiveProfile == null) return;
-
-            IsBusy = true;
-            // Distinguishes a deliberate user disconnect (no auto-reconnect) from an
-            // unexpected service-side loss (auto-reconnect if enabled).
-            _userInitiatedDisconnect = true;
-            try
-            {
-                await _connectionManager.DisconnectProfileAsync(ActiveProfile);
-            }
-            catch (Exception ex)
-            {
-                StatusMessage = $"Disconnect error: {ex.Message}";
-                _logger.LogError(ex, "Error disconnecting profile {Name}", ActiveProfile.Name);
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        /// <summary>
-        /// Starts the auto-reconnect loop if the AutoReconnect preference is enabled.
-        /// Retries ConnectProfileAsync on the configured interval until it succeeds, the
-        /// profile changes, the user disconnects, or the view model is disposed.
-        /// </summary>
-        private void StartAutoReconnectIfEnabled(string reason)
-        {
-            if (_disposed)
-                return;
-
-            if (_settingsService is not { AutoReconnect: true })
-                return;
-
-            lock (_autoReconnectLock)
-            {
-                if (_autoReconnectCts != null)
-                    return; // a loop is already running
-
-                _autoReconnectCts = new CancellationTokenSource();
-            }
-
-            _logger.LogInformation("Auto-reconnect started ({Reason})", reason);
-            _ = AutoReconnectLoopAsync(_autoReconnectCts.Token);
-        }
-
-        private void StopAutoReconnect()
-        {
-            lock (_autoReconnectLock)
-            {
-                _autoReconnectCts?.Cancel();
-                // Disposal happens in the loop's finally.
-            }
-        }
-
-        private async Task AutoReconnectLoopAsync(CancellationToken token)
-        {
-            try
-            {
-                while (!token.IsCancellationRequested && !_disposed)
-                {
-                    // Read the interval each iteration so preference changes apply without a restart.
-                    var intervalMs = Math.Max(100, _settingsService?.AutoReconnectIntervalMs ?? 5000);
-                    try
-                    {
-                        await Task.Delay(intervalMs, token);
-                    }
-                    catch (OperationCanceledException)
-                    {
-                        return;
-                    }
-
-                    var profile = ActiveProfile;
-                    if (profile is null || profile.IsConnected)
-                        return;
-
-                    await _dispatcher.InvokeAsync(() => StatusMessage = $"Connection lost. Retrying in {intervalMs / 1000.0:0.#} s...");
-
-                    bool success;
-                    try
-                    {
-                        success = await _connectionManager.ConnectProfileAsync(profile);
-                    }
-                    catch (Exception ex) when (ex is not (OutOfMemoryException or OperationCanceledException))
-                    {
-                        _logger.LogWarning(ex, "Auto-reconnect attempt for profile {Name} threw", profile.Name);
-                        success = false;
-                    }
-
-                    if (token.IsCancellationRequested || _disposed)
-                        return;
-
-                    if (success)
-                    {
-                        // ProfileConnected (marshaled) stops the loop; also break here in case it raced.
-                        _logger.LogInformation("Auto-reconnect succeeded for profile {Name}", profile.Name);
-                        return;
-                    }
-
-                    _logger.LogWarning("Auto-reconnect attempt failed for profile {Name}; retrying in {Interval} ms", profile.Name, intervalMs);
-                }
-            }
-            catch (OperationCanceledException)
-            {
-            }
-            catch (Exception ex) when (ex is not OutOfMemoryException)
-            {
-                _logger.LogError(ex, "Auto-reconnect loop failed");
-            }
-            finally
-            {
-                lock (_autoReconnectLock)
-                {
-                    if (_autoReconnectCts != null && ReferenceEquals(_autoReconnectCts.Token, token))
-                    {
-                        _autoReconnectCts.Dispose();
-                        _autoReconnectCts = null;
-                    }
-                }
-            }
-        }
-
-        private async Task ToggleConnectionAsync()
-        {
-            _logger.LogInformation("ToggleConnectionAsync invoked");
-            if (IsConnected)
-                await DisconnectAsync();
-            else
-                await ConnectAsync();
-        }
-
-        private async Task ReadAsync()
-        {
-            if (ActiveProfile == null || ActiveService == null) return;
-            await ReadAreaWithBusyAsync(SelectedArea);
-        }
-
-        private async Task ReadHoldingRegistersAsync()
-        {
-            if (ActiveProfile == null || ActiveService == null) return;
-            SelectedArea = PlcArea.HoldingRegister;
-            await ReadAreaWithBusyAsync(PlcArea.HoldingRegister);
-        }
-
-        private async Task ReadInputRegistersAsync()
-        {
-            if (ActiveProfile == null || ActiveService == null) return;
-            SelectedArea = PlcArea.InputRegister;
-            await ReadAreaWithBusyAsync(PlcArea.InputRegister);
-        }
-
-        private async Task ReadCoilsAsync()
-        {
-            if (ActiveProfile == null || ActiveService == null) return;
-            SelectedArea = PlcArea.Coil;
-            await ReadAreaWithBusyAsync(PlcArea.Coil);
-        }
-
-        private async Task ReadDiscreteInputsAsync()
-        {
-            if (ActiveProfile == null || ActiveService == null) return;
-            SelectedArea = PlcArea.DiscreteInput;
-            await ReadAreaWithBusyAsync(PlcArea.DiscreteInput);
-        }
-
-        private async Task ReadAreaWithBusyAsync(PlcArea area)
-        {
-            IsBusy = true;
-            try
-            {
-                await Task.Run(() => ReadAreaAsync(area, CancellationToken.None));
-            }
-            catch (Exception ex)
-            {
-                _dispatcher.Invoke(() => StatusMessage = $"Read error: {ex.Message}");
-                _logger.LogError(ex, "Manual read failed");
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        private async Task WriteAsync() => await PromptAndWriteAsync(SelectedArea);
-
-        private async Task WriteHoldingRegisterAsync() => await PromptAndWriteAsync(PlcArea.HoldingRegister);
-
-        private async Task WriteCoilAsync() => await PromptAndWriteAsync(PlcArea.Coil);
-
-        private async Task PromptAndWriteAsync(PlcArea area)
-        {
-            if (ActiveProfile == null || ActiveService == null || _inputDialogService == null) return;
-            if (area is not (PlcArea.HoldingRegister or PlcArea.Coil)) return;
-
-            var address = PromptAddress($"Write {area}", GetAreaStart(area));
-            if (!address.HasValue) return;
-
-            try
-            {
-                IsBusy = true;
-                await WriteValueToAreaAsync(area, address.Value);
-                await ReadAreaAsync(area, CancellationToken.None);
-            }
-            catch (Exception ex) when (ex is not (OutOfMemoryException or OperationCanceledException))
-            {
-                _logger.LogError(ex, "Write failed");
-                _dispatcher.Invoke(() => StatusMessage = $"Write error: {ex.Message}");
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        private async Task WriteValueToAreaAsync(PlcArea area, int address)
-        {
-            if (_inputDialogService == null || ActiveService == null)
-                return;
-
-            var service = ActiveService;
-            if (service == null) return;
-
-            await _modbusIoGate.WaitAsync(CancellationToken.None);
-            try
-            {
-                var unitId = EffectiveUnitId;
-
-                if (area == PlcArea.HoldingRegister)
-                {
-                    var valueText = _inputDialogService.TryGetInput("Write Value", "Value:", "0", out var input) ? input : null;
-                    if (string.IsNullOrWhiteSpace(valueText) || !ushort.TryParse(valueText, out var value))
-                    {
-                        _dispatcher.Invoke(() => StatusMessage = "Invalid register value.");
-                        return;
-                    }
-
-                    await service.WriteSingleRegisterAsync(unitId, address, value);
-                    _dispatcher.Invoke(() => StatusMessage = $"Wrote {value} to holding register {address}.");
-                }
-                else if (area == PlcArea.Coil)
-                {
-                    var valueText = _inputDialogService.TryGetInput("Write Coil", "Value (true/false):", "false", out var input) ? input : null;
-                    if (string.IsNullOrWhiteSpace(valueText) || !TryParseBool(valueText, out var value))
-                    {
-                        _dispatcher.Invoke(() => StatusMessage = "Invalid coil value. Use true/false, 1/0, on/off.");
-                        return;
-                    }
-
-                    await service.WriteSingleCoilAsync(unitId, address, value);
-                    _dispatcher.Invoke(() => StatusMessage = $"Wrote {value} to coil {address}.");
-                }
-            }
-            finally
-            {
-                _modbusIoGate.Release();
-            }
-        }
-
-        public async Task WriteHoldingRegisterFromEditAsync(RegisterEntry? entry)
-        {
-            if (entry == null || ActiveProfile == null || ActiveService == null) return;
-
-            IsBusy = true;
-            try
-            {
-                var unitId = EffectiveUnitId;
-                var type = (entry.Type ?? "int").ToLowerInvariant();
-                var text = (entry.ValueText ?? string.Empty).Trim().Replace(',', '.');
-
-                await _modbusIoGate.WaitAsync(CancellationToken.None);
-                try
-                {
-                    switch (type)
-                    {
-                        case "real":
-                            if (float.TryParse(text, NumberStyles.Float, CultureInfo.InvariantCulture, out float f))
-                            {
-                                var words = DataTypeConverter.ToUInt16(f, RegistersSwapBytes, RegistersSwapWords);
-                                await ActiveService.WriteRegistersAsync(unitId, entry.Address, words);
-                            }
-                            else
-                            {
-                                _dispatcher.Invoke(() => StatusMessage = $"Invalid float value: {entry.ValueText}");
-                                return;
-                            }
-                            break;
-
-                        case "string":
-                            // A string write expands to one register per two characters. Without a cap a
-                            // long string would silently clobber every register after the target address
-                            // (or fail deep in NModbus past the FC16 limit with a cryptic error).
-                            var stringWords = DataTypeConverter.ToUInt16(text);
-                            if (stringWords.Length > ModbusAddressValidator.MaxWriteRegisters)
-                            {
-                                _dispatcher.Invoke(() => StatusMessage =
-                                    $"String write too large: {stringWords.Length} registers (max {ModbusAddressValidator.MaxWriteRegisters}).");
-                                return;
-                            }
-                            await ActiveService.WriteRegistersAsync(unitId, entry.Address, stringWords);
-                            break;
-
-                        case "int":
-                            if (int.TryParse(text, NumberStyles.Integer, CultureInfo.InvariantCulture, out int iv))
-                            {
-                                await ActiveService.WriteSingleRegisterAsync(unitId, entry.Address, unchecked((ushort)iv));
-                            }
-                            else
-                            {
-                                _dispatcher.Invoke(() => StatusMessage = $"Invalid integer value: {entry.ValueText}");
-                                return;
-                            }
-                            break;
-
-                        default:
-                            if (uint.TryParse(text, NumberStyles.Integer, CultureInfo.InvariantCulture, out uint uv) && uv <= ushort.MaxValue)
-                            {
-                                await ActiveService.WriteSingleRegisterAsync(unitId, entry.Address, (ushort)uv);
-                            }
-                            else
-                            {
-                                _dispatcher.Invoke(() => StatusMessage = $"Invalid unsigned value: {entry.ValueText}");
-                                return;
-                            }
-                            break;
-                    }
-                }
-                finally
-                {
-                    _modbusIoGate.Release();
-                }
-
-                await ReadAreaAsync(PlcArea.HoldingRegister, CancellationToken.None);
-            }
-            catch (Exception ex) when (ex is not (OutOfMemoryException or OperationCanceledException))
-            {
-                _logger.LogError(ex, "Holding register write failed");
-                _dispatcher.Invoke(() => StatusMessage = $"Write error: {ex.Message}");
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        public async Task WriteCoilFromEditAsync(CoilEntry? entry)
-        {
-            if (entry == null || ActiveProfile == null || ActiveService == null) return;
-
-            IsBusy = true;
-            try
-            {
-                await _modbusIoGate.WaitAsync(CancellationToken.None);
-                try
-                {
-                    await ActiveService.WriteSingleCoilAsync(EffectiveUnitId, entry.Address, entry.State);
-                }
-                finally
-                {
-                    _modbusIoGate.Release();
-                }
-
-                await ReadAreaAsync(PlcArea.Coil, CancellationToken.None);
-            }
-            catch (Exception ex) when (ex is not (OutOfMemoryException or OperationCanceledException))
-            {
-                _logger.LogError(ex, "Coil write failed");
-                _dispatcher.Invoke(() => StatusMessage = $"Write error: {ex.Message}");
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        private int? PromptAddress(string title, int defaultAddress)
-        {
-            if (_inputDialogService == null) return null;
-
-            var defaultValue = defaultAddress.ToString(CultureInfo.InvariantCulture);
-            if (!_inputDialogService.TryGetInput(title, "Address:", defaultValue, out var input) ||
-                !int.TryParse(input, out var address))
-            {
-                _dispatcher.Invoke(() => StatusMessage = "Invalid address.");
-                return null;
-            }
-
-            if (address is < ModbusAddressValidator.MinStartAddress or > ModbusAddressValidator.MaxStartAddress)
-            {
-                _dispatcher.Invoke(() => StatusMessage =
-                    $"Address must be between {ModbusAddressValidator.MinStartAddress} and {ModbusAddressValidator.MaxStartAddress}.");
-                return null;
-            }
-
-            return address;
-        }
-
-        private void StartPolling()
-        {
-            if (_disposed || ActiveProfile is not { IsConnected: true } || !AnyMonitorEnabled()) return;
-
-            lock (_pollLifecycleLock)
-            {
-                if (_pollCts != null) return;
-
-                var cts = new CancellationTokenSource();
-                _pollCts = cts;
-                _ = Task.Run(() => PollLoopAsync(cts), cts.Token);
-            }
-        }
-
-        private void StopPolling()
-        {
-            lock (_pollLifecycleLock)
-            {
-                _pollCts?.Cancel();
-            }
-
-            lock (_pendingPollLock)
-            {
-                _pendingPollAreas.Clear();
-            }
-        }
-
-        private async Task PollLoopAsync(CancellationTokenSource loopCts)
-        {
-            var token = loopCts.Token;
-            var hadError = false;
-            try
-            {
-                while (!token.IsCancellationRequested && ActiveProfile is { IsConnected: true } && AnyMonitorEnabled())
-                {
-                    var now = DateTime.UtcNow;
-                    QueueDueAreaReads(now);
-                    await DrainPendingAreaReadsAsync(token);
-                    await Task.Delay(PollLoopIntervalMs, token);
-                }
-            }
-            catch (OperationCanceledException)
-            {
-                _logger.LogDebug("Avalonia polling loop canceled");
-            }
-            catch (Exception ex) when (ex is not OutOfMemoryException)
-            {
-                hadError = true;
-                _logger.LogError(ex, "Avalonia polling loop failed");
-                await _dispatcher.InvokeAsync(() => StatusMessage = $"Poll error: {ex.Message}");
-            }
-            finally
-            {
-                var restart = false;
-                lock (_pollLifecycleLock)
-                {
-                    if (ReferenceEquals(_pollCts, loopCts))
-                    {
-                        _pollCts = null;
-                        restart = !_disposed && ActiveProfile is { IsConnected: true } && AnyMonitorEnabled();
-                    }
-                }
-
-                loopCts.Dispose();
-                if (restart)
-                {
-                    // A clean stop (disconnected, monitors disabled) restarts immediately;
-                    // an error exit backs off first so a persistently failing service
-                    // cannot spin a tight error → restart loop.
-                    if (hadError)
-                        RestartPollingAfterBackoff();
-                    else
-                        StartPolling();
-                }
-            }
-        }
-
-        private void RestartPollingAfterBackoff()
-        {
-            _ = Task.Run(async () =>
-            {
-                try
-                {
-                    await Task.Delay(PollRestartBackoffMs);
-                }
-                catch (OperationCanceledException)
-                {
-                    return;
-                }
-
-                // Re-evaluate after the pause: the connection may have dropped (or the
-                // view model been disposed) in the meantime.
-                if (!_disposed && ActiveProfile is { IsConnected: true } && AnyMonitorEnabled())
-                {
-                    StartPolling();
-                }
-            });
-        }
-
-        private void QueueDueAreaReads(DateTime now)
-        {
-            QueueAreaReadIfDue(PlcArea.HoldingRegister, HoldingMonitorEnabled, HoldingMonitorPeriodMs, _lastHoldingReadUtc, now, value => _lastHoldingReadUtc = value);
-            QueueAreaReadIfDue(PlcArea.InputRegister, InputRegistersMonitorEnabled, InputRegistersMonitorPeriodMs, _lastInputRegReadUtc, now, value => _lastInputRegReadUtc = value);
-            QueueAreaReadIfDue(PlcArea.Coil, CoilsMonitorEnabled, CoilsMonitorPeriodMs, _lastCoilsReadUtc, now, value => _lastCoilsReadUtc = value);
-            QueueAreaReadIfDue(PlcArea.DiscreteInput, DiscreteInputsMonitorEnabled, DiscreteInputsMonitorPeriodMs, _lastDiscreteReadUtc, now, value => _lastDiscreteReadUtc = value);
-        }
-
-        private void QueueAreaReadIfDue(
-            PlcArea area,
-            bool enabled,
-            int periodMs,
-            DateTime lastReadUtc,
-            DateTime now,
-            Action<DateTime> setLastReadUtc)
-        {
-            if (!enabled || (now - lastReadUtc).TotalMilliseconds < Math.Max(MinimumMonitorPeriodMs, periodMs))
-                return;
-
-            lock (_pendingPollLock)
-            {
-                _pendingPollAreas.Add(area);
-            }
-
-            setLastReadUtc(now);
-        }
-
-        private async Task DrainPendingAreaReadsAsync(CancellationToken token)
-        {
-            PlcArea[] pending;
-            lock (_pendingPollLock)
-            {
-                pending = _pendingPollAreas.ToArray();
-                _pendingPollAreas.Clear();
-            }
-
-            foreach (var area in pending)
-            {
-                token.ThrowIfCancellationRequested();
-                if (IsAreaMonitorEnabled(area))
-                {
-                    await ReadAreaAsync(area, token, true);
-                }
-            }
-        }
-
-        private async Task ReadAreaAsync(PlcArea area, CancellationToken token, bool isMonitoring = false)
-        {
-            await _modbusIoGate.WaitAsync(token);
-            try
-            {
-                var service = ActiveService;
-                if (service == null || ActiveProfile == null)
-                {
-                    await _dispatcher.InvokeAsync(() => StatusMessage = "No active service.");
-                    return;
-                }
-
-                var unitId = EffectiveUnitId;
-                var (start, count) = GetAreaStartCount(area);
-
-                await _dispatcher.InvokeAsync(() => StatusMessage = $"Reading {area}...");
-                token.ThrowIfCancellationRequested();
-
-                switch (area)
-                {
-                    case PlcArea.HoldingRegister:
-                        var holding = await service.ReadHoldingRegistersAsync(unitId, start, count)
-                            ?? throw new InvalidOperationException("Read returned no response.");
-                        var holdingPartial = holding.Length < count;
-                        await _dispatcher.InvokeAsync(() =>
-                        {
-                            if (IsRegisterGridEditing) return;
-                            HoldingRegisters = ApplyRegisterValues(
-                                HoldingRegisters,
-                                start,
-                                holding,
-                                RegistersGlobalType,
-                                RegistersSwapBytes,
-                                RegistersSwapWords,
-                                CurrentConfig.RegisterSettings.HoldingRegisterMetadata,
-                                holdingPartial);
-                            StatusMessage = holdingPartial
-                                ? $"Partial read: {holding.Length} of {count} holding registers"
-                                : $"Read {holding.Length} holding registers";
-                        });
-                        break;
-
-                    case PlcArea.InputRegister:
-                        var input = await service.ReadInputRegistersAsync(unitId, start, count)
-                            ?? throw new InvalidOperationException("Read returned no response.");
-                        var inputPartial = input.Length < count;
-                        await _dispatcher.InvokeAsync(() =>
-                        {
-                            if (IsRegisterGridEditing) return;
-                            InputRegisters = ApplyRegisterValues(
-                                InputRegisters,
-                                start,
-                                input,
-                                InputRegistersGlobalType,
-                                InputRegistersSwapBytes,
-                                InputRegistersSwapWords,
-                                CurrentConfig.RegisterSettings.InputRegisterMetadata,
-                                inputPartial);
-                            StatusMessage = inputPartial
-                                ? $"Partial read: {input.Length} of {count} input registers"
-                                : $"Read {input.Length} input registers";
-                        });
-                        break;
-
-                    case PlcArea.Coil:
-                        var coils = await service.ReadCoilsAsync(unitId, start, count)
-                            ?? throw new InvalidOperationException("Read returned no response.");
-                        var coilsPartial = coils.Length < count;
-                        await _dispatcher.InvokeAsync(() =>
-                        {
-                            if (IsRegisterGridEditing) return;
-                            Coils = ApplyCoilValues(Coils, start, coils, coilsPartial);
-                            StatusMessage = coilsPartial
-                                ? $"Partial read: {coils.Length} of {count} coils"
-                                : $"Read {coils.Length} coils";
-                        });
-                        break;
-
-                    case PlcArea.DiscreteInput:
-                        var discrete = await service.ReadDiscreteInputsAsync(unitId, start, count)
-                            ?? throw new InvalidOperationException("Read returned no response.");
-                        var discretePartial = discrete.Length < count;
-                        await _dispatcher.InvokeAsync(() =>
-                        {
-                            if (IsRegisterGridEditing) return;
-                            DiscreteInputs = ApplyCoilValues(DiscreteInputs, start, discrete, discretePartial);
-                            StatusMessage = discretePartial
-                                ? $"Partial read: {discrete.Length} of {count} discrete inputs"
-                                : $"Read {discrete.Length} discrete inputs";
-                        });
-                        break;
-                }
-
-                ClearMonitorFailure(area);
-            }
-            catch (OperationCanceledException)
-            {
-                throw;
-            }
-            catch (Exception ex) when (ex is not OutOfMemoryException)
-            {
-                await HandleAreaReadFailureAsync(area, ex, isMonitoring);
-            }
-            finally
-            {
-                _modbusIoGate.Release();
-            }
-        }
-
-        private bool IsAreaMonitorEnabled(PlcArea area) => area switch
-        {
-            PlcArea.HoldingRegister => HoldingMonitorEnabled,
-            PlcArea.InputRegister => InputRegistersMonitorEnabled,
-            PlcArea.Coil => CoilsMonitorEnabled,
-            PlcArea.DiscreteInput => DiscreteInputsMonitorEnabled,
-            _ => false
-        };
-
-        private async Task HandleAreaReadFailureAsync(PlcArea area, Exception exception, bool isMonitoring)
-        {
-            var failureTime = DateTime.UtcNow;
-            int failureCount;
-            lock (_monitorFailureLock)
-            {
-                failureCount = GetMonitorFailureCountUnsafe(area) + 1;
-                _monitorFailureCounts[area] = failureCount;
-                _lastMonitorFailureUtc[area] = failureTime;
-            }
-
-            // Runs on the poll thread - marshal observable property changes to the UI thread.
-            await _dispatcher.InvokeAsync(() =>
-            {
-                LastErrorTime = failureTime;
-                HasConnectionError = true;
-            });
-            _logger.LogError(exception, "Error reading {Area} (failure {FailureCount})", area, failureCount);
-
-            var message = $"Failed to read {area}: {exception.Message}";
-            if (isMonitoring && IsAreaMonitorEnabled(area))
-            {
-                message += "\n\nContinuous monitoring has been paused. Fix the issue and re-enable monitoring.";
-                await _dispatcher.InvokeAsync(() => SetAreaMonitorEnabled(area, false));
-            }
-
-            await _dispatcher.InvokeAsync(() => StatusMessage = message);
-
-            if (_messageBoxService != null)
-            {
-                Task<DialogResult>? dialogTask = null;
-                await _dispatcher.InvokeAsync(() =>
-                    dialogTask = _messageBoxService.ShowAsync(message, "Read Error", DialogButton.Ok, DialogIcon.Error));
-                if (dialogTask != null)
-                {
-                    await dialogTask;
-                }
-            }
-        }
-
-        private int GetMonitorFailureCountUnsafe(PlcArea area) =>
-            _monitorFailureCounts.TryGetValue(area, out var count) ? count : 0;
-
-        private void ClearMonitorFailure(PlcArea area)
-        {
-            var allClear = false;
-            lock (_monitorFailureLock)
-            {
-                _monitorFailureCounts[area] = 0;
-                _lastMonitorFailureUtc.Remove(area);
-                allClear = _monitorFailureCounts.Values.All(count => count == 0);
-            }
-
-            if (allClear)
-            {
-                // May run on the poll thread - marshal the observable property change to the UI thread.
-                _dispatcher.Invoke(() => HasConnectionError = false);
-            }
-        }
-
-        private void ResetMonitorFailures()
-        {
-            lock (_monitorFailureLock)
-            {
-                _monitorFailureCounts.Clear();
-                _lastMonitorFailureUtc.Clear();
-            }
-
-            LastErrorTime = DateTime.MinValue;
-        }
-
-        private void SetAreaMonitorEnabled(PlcArea area, bool enabled)
-        {
-            switch (area)
-            {
-                case PlcArea.HoldingRegister:
-                    HoldingMonitorEnabled = enabled;
-                    break;
-                case PlcArea.InputRegister:
-                    InputRegistersMonitorEnabled = enabled;
-                    break;
-                case PlcArea.Coil:
-                    CoilsMonitorEnabled = enabled;
-                    break;
-                case PlcArea.DiscreteInput:
-                    DiscreteInputsMonitorEnabled = enabled;
-                    break;
-            }
-        }
-
-        private ObservableCollection<RegisterEntry> ApplyRegisterValues(
-            ObservableCollection<RegisterEntry>? target,
-            int start,
-            ushort[] values,
-            string globalType,
-            bool swapBytes,
-            bool swapWords,
-            IEnumerable<RegisterMetadata>? metadata = null,
-            bool isPartialRead = false)
-        {
-            target ??= new ObservableCollection<RegisterEntry>();
-            var metadataByAddress = metadata?.GroupBy(item => item.Address)
-                                    .ToDictionary(g => g.Key, g => g.First())
-                                    ?? new Dictionary<int, RegisterMetadata>();
-
-            // Duplicate addresses in the grid (user editing) must not crash the read path;
-            // the first entry wins and is the one updated below.
-            var entriesByAddress = new Dictionary<int, RegisterEntry>();
-            foreach (var existing in target)
-            {
-                entriesByAddress.TryAdd(existing.Address, existing);
-            }
-            var usedAddresses = new HashSet<int>();
-
-            int idx = 0;
-            while (idx < values.Length)
-            {
-                var address = start + idx;
-                var savedMetadata = metadataByAddress.GetValueOrDefault(address);
-                var type = (savedMetadata?.Type ?? globalType).ToLowerInvariant();
-                usedAddresses.Add(address);
-
-                if (!entriesByAddress.TryGetValue(address, out var entry))
-                {
-                    entry = new RegisterEntry();
-                    target.Add(entry);
-                    entriesByAddress[address] = entry;
-                }
-
-                entry.Address = address;
-                entry.Value = values[idx];
-                entry.Type = type;
-                entry.SwapBytes = savedMetadata?.SwapBytes ?? swapBytes;
-                entry.SwapWords = savedMetadata?.SwapWords ?? swapWords;
-                entry.IsReadError = false;
-                entry.ReadErrorMessage = null;
-
-                switch (type)
-                {
-                    case "int":
-                        entry.ValueText = unchecked((short)values[idx]).ToString(CultureInfo.InvariantCulture);
-                        idx += 1;
-                        break;
-
-                    case "real":
-                        if (idx + 1 < values.Length)
-                        {
-                            entry.ValueText = DataTypeConverter.ToSingle(values[idx], values[idx + 1], entry.SwapBytes, entry.SwapWords).ToString(CultureInfo.InvariantCulture);
-
-                            var nextAddress = address + 1;
-                            usedAddresses.Add(nextAddress);
-                            if (!entriesByAddress.TryGetValue(nextAddress, out var next))
-                            {
-                                next = new RegisterEntry();
-                                target.Add(next);
-                                entriesByAddress[nextAddress] = next;
-                            }
-
-                            next.Address = nextAddress;
-                            next.Value = values[idx + 1];
-                            next.Type = type;
-                            next.SwapBytes = swapBytes;
-                            next.SwapWords = swapWords;
-                            next.ValueText = string.Empty;
-                            next.IsReadError = false;
-                            next.ReadErrorMessage = null;
-
-                            idx += 2;
-                        }
-                        else
-                        {
-                            entry.ValueText = values[idx].ToString(CultureInfo.InvariantCulture);
-                            idx += 1;
-                        }
-                        break;
-
-                    case "string":
-                        entry.ValueText = DataTypeConverter.ToString(values[idx]);
-                        idx += 1;
-                        break;
-
-                    default:
-                        entry.ValueText = values[idx].ToString(CultureInfo.InvariantCulture);
-                        idx += 1;
-                        break;
-                }
-            }
-
-            for (int i = target.Count - 1; i >= 0; i--)
-            {
-                if (!usedAddresses.Contains(target[i].Address))
-                {
-                    target.RemoveAt(i);
-                }
-            }
-
-            if (isPartialRead)
-            {
-                var lastVisible = target
-                    .Where(e => !string.IsNullOrEmpty(e.ValueText))
-                    .OrderByDescending(e => e.Address)
-                    .FirstOrDefault();
-                if (lastVisible != null)
-                {
-                    lastVisible.IsReadError = true;
-                    lastVisible.ReadErrorMessage = "Read stopped here. The remaining registers could not be read.";
-                }
-            }
-
-            EnsureSortedByAddress(target);
-            return target;
-        }
-
-        private static void EnsureSortedByAddress(ObservableCollection<RegisterEntry> collection)
-        {
-            if (collection.Count < 2) return;
-
-            bool sorted = true;
-            for (int i = 1; i < collection.Count; i++)
-            {
-                if (collection[i].Address < collection[i - 1].Address)
-                {
-                    sorted = false;
-                    break;
-                }
-            }
-
-            if (sorted) return;
-
-            var entries = collection.OrderBy(e => e.Address).ToList();
-            collection.Clear();
-            foreach (var entry in entries)
-            {
-                collection.Add(entry);
-            }
-        }
-
-        private ObservableCollection<CoilEntry> ApplyCoilValues(ObservableCollection<CoilEntry>? target, int start, bool[] values, bool isPartialRead = false)
-        {
-            target ??= new ObservableCollection<CoilEntry>();
-            var entriesByAddress = target.ToDictionary(e => e.Address);
-            var usedAddresses = new HashSet<int>();
-
-            for (int i = 0; i < values.Length; i++)
-            {
-                var address = start + i;
-                usedAddresses.Add(address);
-                if (!entriesByAddress.TryGetValue(address, out var entry))
-                {
-                    entry = new CoilEntry { Address = address };
-                    target.Add(entry);
-                    entriesByAddress[address] = entry;
-                }
-
-                entry.State = values[i];
-                entry.IsReadError = false;
-                entry.ReadErrorMessage = null;
-            }
-
-            for (int i = target.Count - 1; i >= 0; i--)
-            {
-                if (!usedAddresses.Contains(target[i].Address))
-                {
-                    target.RemoveAt(i);
-                }
-            }
-
-            if (isPartialRead && target.Count > 0)
-            {
-                var lastCoil = target.OrderByDescending(e => e.Address).First();
-                lastCoil.IsReadError = true;
-                lastCoil.ReadErrorMessage = "Read stopped here. The remaining coils could not be read.";
-            }
-
-            EnsureSortedByAddress(target);
-            return target;
-        }
-
-        private static void EnsureSortedByAddress(ObservableCollection<CoilEntry> collection)
-        {
-            if (collection.Count < 2) return;
-
-            bool sorted = true;
-            for (int i = 1; i < collection.Count; i++)
-            {
-                if (collection[i].Address < collection[i - 1].Address)
-                {
-                    sorted = false;
-                    break;
-                }
-            }
-
-            if (sorted) return;
-
-            var entries = collection.OrderBy(e => e.Address).ToList();
-            collection.Clear();
-            foreach (var entry in entries)
-            {
-                collection.Add(entry);
-            }
-        }
-
-        #region Custom Watch
-
-        private async Task AddCustomEntryAsync()
-        {
-            await _dispatcher.InvokeAsync(() =>
-            {
-                int nextAddress = 1;
-                string type = "int";
-                string area = "HoldingRegister";
-                string name = "Tag0";
-
-                if (CustomEntries.Count > 0)
-                {
-                    var last = CustomEntries[^1];
-                    type = last.Type ?? "int";
-                    area = last.Area ?? "HoldingRegister";
-                    name = GenerateNextName(last.Name);
-
-                    int increment = IsMultiRegisterType(type)
-                        ? MultiRegisterTypeIncrement
-                        : SingleRegisterTypeIncrement;
-                    nextAddress = Math.Max(1, last.Address + increment);
-                }
-
-                var entry = new CustomEntry
-                {
-                    Name = name,
-                    Address = nextAddress,
-                    Area = area,
-                    Type = type,
-                    Value = "0",
-                    WriteValue = "0",
-                    Continuous = false,
-                    PeriodMs = DefaultCustomPeriodMs,
-                    Monitor = false,
-                    ReadPeriodMs = DefaultCustomPeriodMs,
-                    Trend = false
-                };
-
-                CustomEntries.Add(entry);
-                SelectedCustomEntry = entry;
-                ReadAllCustomNowCommand.NotifyCanExecuteChanged();
-                StatusMessage = $"Added custom entry {name}.";
-            });
-        }
-
-        private Task AddCustomBulkEntryAsync()
-        {
-            if (_customBulkAddDialogService == null) return Task.CompletedTask;
-
-            return _dispatcher.InvokeAsync(() =>
-            {
-                if (!_customBulkAddDialogService.TryGetBulkAdd(out var result) || result == null)
-                {
-                    StatusMessage = "Bulk add cancelled.";
-                    return;
-                }
-
-                int increment = IsMultiRegisterType(result.Type)
-                    ? MultiRegisterTypeIncrement
-                    : SingleRegisterTypeIncrement;
-
-                for (int i = 0; i < result.Count; i++)
-                {
-                    int address = result.StartRegister + i * increment;
-                    var entry = new CustomEntry
-                    {
-                        Name = $"{result.NamePrefix}{i}",
-                        Address = address,
-                        Area = result.Area,
-                        Type = result.Type,
-                        Value = "0",
-                        WriteValue = "0",
-                        Continuous = false,
-                        PeriodMs = result.WritePeriodMs,
-                        Monitor = false,
-                        ReadPeriodMs = result.ReadPeriodMs,
-                        Trend = false
-                    };
-
-                    CustomEntries.Add(entry);
-                }
-
-                SelectedCustomEntry = CustomEntries[^1];
-                ReadAllCustomNowCommand.NotifyCanExecuteChanged();
-                StatusMessage = $"Added {result.Count} custom entries starting at {result.StartRegister}.";
-            });
-        }
-
-        private static bool IsMultiRegisterType(string type) =>
-            type.Equals("real", StringComparison.OrdinalIgnoreCase) ||
-            type.Equals("float", StringComparison.OrdinalIgnoreCase) ||
-            type.Equals("dword", StringComparison.OrdinalIgnoreCase) ||
-            type.Equals("dint", StringComparison.OrdinalIgnoreCase);
-
-        private async Task RemoveCustomEntryAsync()
-        {
-            if (SelectedCustomEntry == null) return;
-
-            await _dispatcher.InvokeAsync(() =>
-            {
-                CustomEntries.Remove(SelectedCustomEntry);
-                SelectedCustomEntry = null;
-                ReadAllCustomNowCommand.NotifyCanExecuteChanged();
-                StatusMessage = "Removed custom entry.";
-            });
-        }
-
-        private Task DeleteCustomEntryAsync(CustomEntry? entry)
-        {
-            if (entry == null) return Task.CompletedTask;
-
-            return _dispatcher.InvokeAsync(() =>
-            {
-                CustomEntries.Remove(entry);
-                if (SelectedCustomEntry == entry)
-                {
-                    SelectedCustomEntry = null;
-                }
-
-                ReadAllCustomNowCommand.NotifyCanExecuteChanged();
-                StatusMessage = $"Removed custom entry {entry.Name}.";
-            });
-        }
-
-        private Task ReadSelectedCustomEntryAsync() => ReadCustomEntryNowAsync(SelectedCustomEntry);
-
-        private Task WriteSelectedCustomEntryAsync() => WriteCustomEntryNowAsync(SelectedCustomEntry);
-
-        private async Task ReadCustomEntryNowAsync(CustomEntry? entry)
-        {
-            if (entry == null || ActiveService == null || ActiveProfile == null) return;
-
-            IsBusy = true;
-            try
-            {
-                var value = await ReadCustomValueSerializedAsync(entry, CancellationToken.None);
-                var readAt = DateTime.UtcNow;
-                await _dispatcher.InvokeAsync(() =>
-                {
-                    entry.Value = value;
-                    entry.LastReadUtc = readAt;
-                    StatusMessage = $"Read {entry.Name} = {value}";
-                });
-            }
-            catch (OperationCanceledException)
-            {
-                throw;
-            }
-            catch (Exception ex) when (ex is not OutOfMemoryException)
-            {
-                await ReportCustomOperationFailureAsync("read", entry, ex, false);
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        private async Task WriteCustomEntryNowAsync(CustomEntry? entry)
-        {
-            if (entry == null || ActiveService == null || ActiveProfile == null) return;
-
-            IsBusy = true;
-            try
-            {
-                var result = await WriteCustomValueSerializedAsync(entry, CancellationToken.None);
-                var writtenAt = DateTime.UtcNow;
-                await _dispatcher.InvokeAsync(() =>
-                {
-                    if (result)
-                    {
-                        entry.LastWriteUtc = writtenAt;
-                    }
-
-                    StatusMessage = result
-                        ? $"Wrote {entry.Name}"
-                        : $"Write completed with issues for {entry.Name}.";
-                });
-            }
-            catch (OperationCanceledException)
-            {
-                throw;
-            }
-            catch (Exception ex) when (ex is not OutOfMemoryException)
-            {
-                await ReportCustomOperationFailureAsync("write", entry, ex, false);
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        private async Task ReadAllCustomEntriesAsync()
-        {
-            if (ActiveService == null || ActiveProfile == null || CustomEntries.Count == 0) return;
-
-            IsBusy = true;
-            try
-            {
-                var entries = await _dispatcher.InvokeAsync(() => CustomEntries.ToList());
-                var readCount = 0;
-                foreach (var entry in entries)
-                {
-                    try
-                    {
-                        var value = await ReadCustomValueSerializedAsync(entry, CancellationToken.None);
-                        var readAt = DateTime.UtcNow;
-                        await _dispatcher.InvokeAsync(() =>
-                        {
-                            entry.Value = value;
-                            entry.LastReadUtc = readAt;
-                        });
-                        readCount++;
-                    }
-                    catch (Exception ex) when (ex is not (OutOfMemoryException or OperationCanceledException))
-                    {
-                        await ReportCustomOperationFailureAsync("read", entry, ex, false);
-                    }
-                }
-
-                await _dispatcher.InvokeAsync(() => StatusMessage = $"Read {readCount} of {entries.Count} custom entries.");
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        private async Task<string> ReadCustomValueSerializedAsync(CustomEntry entry, CancellationToken token)
-        {
-            await _modbusIoGate.WaitAsync(token);
-            try
-            {
-                return await ReadCustomValueAsync(entry);
-            }
-            finally
-            {
-                _modbusIoGate.Release();
-            }
-        }
-
-        private async Task<bool> WriteCustomValueSerializedAsync(CustomEntry entry, CancellationToken token)
-        {
-            await _modbusIoGate.WaitAsync(token);
-            try
-            {
-                return await WriteCustomValueAsync(entry);
-            }
-            finally
-            {
-                _modbusIoGate.Release();
-            }
-        }
-
-        private async Task ReportCustomOperationFailureAsync(string operation, CustomEntry entry, Exception exception, bool monitoring)
-        {
-            _logger.LogError(exception, "Error {Operation} custom entry {Name}", operation, entry.Name);
-            var suffix = monitoring ? " Continuous monitoring has been paused." : string.Empty;
-            var message = $"Failed to {operation} custom entry '{entry.Name}': {exception.Message}.{suffix}";
-            await _dispatcher.InvokeAsync(() => StatusMessage = message);
-            if (_messageBoxService != null)
-            {
-                Task<DialogResult>? dialogTask = null;
-                await _dispatcher.InvokeAsync(() =>
-                    dialogTask = _messageBoxService.ShowAsync(message, "Custom Watch Error", DialogButton.Ok, DialogIcon.Error));
-                if (dialogTask != null)
-                {
-                    await dialogTask;
-                }
-            }
-        }
-
-        private async Task<string> ReadCustomValueAsync(CustomEntry entry)
-        {
-            var service = ActiveService;
-            if (service == null || ActiveProfile == null)
-                throw new InvalidOperationException("No active service.");
-
-            var unitId = EffectiveUnitId;
-            var area = (entry.Area ?? "HoldingRegister").ToLowerInvariant();
-            var type = (entry.Type ?? "int").ToLowerInvariant();
-
-            switch (area)
-            {
-                case "holdingregister":
-                case "inputregister":
-                    int count = type == "real" ? 2 : 1;
-                    var areaEnum = area == "holdingregister" ? PlcArea.HoldingRegister : PlcArea.InputRegister;
-                    var values = areaEnum == PlcArea.HoldingRegister
-                        ? await service.ReadHoldingRegistersAsync(unitId, entry.Address, count)
-                        : await service.ReadInputRegistersAsync(unitId, entry.Address, count);
-
-                    if (values == null || values.Length == 0)
-                        throw new InvalidOperationException("Read returned no response.");
-
-                    if (type == "real" && values.Length < 2)
-                        throw new InvalidOperationException("A REAL value requires two registers.");
-
-                    return type switch
-                    {
-                        "int" => unchecked((short)values[0]).ToString(CultureInfo.InvariantCulture),
-                        "real" => DataTypeConverter.ToSingle(values[0], values[1], false, false).ToString(CultureInfo.InvariantCulture),
-                        "string" => DataTypeConverter.ToString(values[0]),
-                        _ => values[0].ToString(CultureInfo.InvariantCulture)
-                    };
-
-                case "coil":
-                case "discreteinput":
-                    var coilValues = area == "coil"
-                        ? await service.ReadCoilsAsync(unitId, entry.Address, 1)
-                        : await service.ReadDiscreteInputsAsync(unitId, entry.Address, 1);
-                    if (coilValues == null || coilValues.Length == 0) return "No response";
-                    return coilValues[0] ? "1" : "0";
-
-                default:
-                    return $"Unknown area: {entry.Area}";
-            }
-        }
-
-        private async Task<bool> WriteCustomValueAsync(CustomEntry entry)
-        {
-            var service = ActiveService;
-            if (service == null || ActiveProfile == null) return false;
-
-            var unitId = EffectiveUnitId;
-            var area = (entry.Area ?? "HoldingRegister").ToLowerInvariant();
-            var type = (entry.Type ?? "int").ToLowerInvariant();
-
-            switch (area)
-            {
-                case "holdingregister":
-                    switch (type)
-                    {
-                        case "real":
-                            if (float.TryParse(entry.WriteValue, NumberStyles.Float, CultureInfo.InvariantCulture, out float f) ||
-                                float.TryParse(entry.WriteValue, NumberStyles.Float, CultureInfo.CurrentCulture, out f))
-                            {
-                                var words = DataTypeConverter.ToUInt16(f, false, false);
-                                await service.WriteRegistersAsync(unitId, entry.Address, words);
-                                return true;
-                            }
-                            return false;
-
-                        case "string":
-                            // Same FC16 cap as the register grid: reject oversized strings with a clear
-                            // message (surfaced by the caller) instead of clobbering adjacent registers.
-                            var stringWords = DataTypeConverter.ToUInt16(entry.WriteValue ?? string.Empty);
-                            if (stringWords.Length > ModbusAddressValidator.MaxWriteRegisters)
-                            {
-                                throw new ArgumentException(
-                                    $"String write too large: {stringWords.Length} registers (max {ModbusAddressValidator.MaxWriteRegisters}).");
-                            }
-                            await service.WriteRegistersAsync(unitId, entry.Address, stringWords);
-                            return true;
-
-                        case "int":
-                            if (int.TryParse(entry.WriteValue, NumberStyles.Integer, CultureInfo.InvariantCulture, out int iv))
-                            {
-                                await service.WriteSingleRegisterAsync(unitId, entry.Address, unchecked((ushort)iv));
-                                return true;
-                            }
-                            return false;
-
-                        default:
-                            if (uint.TryParse(entry.WriteValue, NumberStyles.Integer, CultureInfo.InvariantCulture, out uint uv))
-                            {
-                                if (uv > 0xFFFF) uv = 0xFFFF;
-                                await service.WriteSingleRegisterAsync(unitId, entry.Address, (ushort)uv);
-                                return true;
-                            }
-                            return false;
-                    }
-
-                case "coil":
-                    if (TryParseBool(entry.WriteValue, out bool b))
-                    {
-                        await service.WriteSingleCoilAsync(unitId, entry.Address, b);
-                        return true;
-                    }
-                    return false;
-
-                default:
-                    return false;
-            }
-        }
-
-        private async Task SaveCustomAsync()
-        {
-            if (_customEntryService == null) return;
-
-            IsBusy = true;
-            try
-            {
-                await _customEntryService.SaveCustomAsync(CustomEntries);
-                StatusMessage = "Custom entries saved.";
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Error saving custom entries");
-                StatusMessage = $"Save error: {ex.Message}";
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        private async Task LoadCustomAsync()
-        {
-            if (_customEntryService == null) return;
-
-            IsBusy = true;
-            try
-            {
-                var entries = await _customEntryService.LoadCustomAsync();
-                if (entries != null)
-                {
-                    await _dispatcher.InvokeAsync(() =>
-                    {
-                        CustomEntries.Clear();
-                        foreach (var e in entries)
-                        {
-                            CustomEntries.Add(e);
-                        }
-                        ReadAllCustomNowCommand.NotifyCanExecuteChanged();
-                    });
-                    StatusMessage = $"Loaded {entries.Count} custom entries.";
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Error loading custom entries");
-                StatusMessage = $"Load error: {ex.Message}";
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        private void StartCustomWatchMonitoring()
-        {
-            if (_disposed || ActiveProfile is not { IsConnected: true } || !IsCustomWatchMonitoring) return;
-
-            lock (_customWatchLifecycleLock)
-            {
-                if (_customWatchCts != null) return;
-
-                var cts = new CancellationTokenSource();
-                _customWatchCts = cts;
-                _ = Task.Run(() => CustomWatchLoopAsync(cts), cts.Token);
-            }
-        }
-
-        private void StopCustomWatchMonitoring()
-        {
-            lock (_customWatchLifecycleLock)
-            {
-                _customWatchCts?.Cancel();
-            }
-        }
-
-        private async Task CustomWatchLoopAsync(CancellationTokenSource loopCts)
-        {
-            var token = loopCts.Token;
-            try
-            {
-                while (!token.IsCancellationRequested && ActiveProfile is { IsConnected: true } && IsCustomWatchMonitoring)
-                {
-                    var entries = await _dispatcher.InvokeAsync(() => CustomEntries.ToList());
-                    var now = DateTime.UtcNow;
-
-                    foreach (var entry in entries)
-                    {
-                        token.ThrowIfCancellationRequested();
-
-                        var readPeriod = entry.ReadPeriodMs <= 0 ? DefaultCustomPeriodMs : entry.ReadPeriodMs;
-                        if (entry.Monitor && (now - entry.LastReadUtc).TotalMilliseconds >= readPeriod)
-                        {
-                            try
-                            {
-                                var value = await ReadCustomValueSerializedAsync(entry, token);
-                                var readAt = DateTime.UtcNow;
-                                await _dispatcher.InvokeAsync(() =>
-                                {
-                                    entry.Value = value;
-                                    entry.LastReadUtc = readAt;
-                                });
-
-                                if (entry.Trend && _trendLogger != null && TryParseTrendValue(value, out var trendValue))
-                                {
-                                    _trendLogger.Publish(entry.Name, trendValue, readAt);
-                                }
-                            }
-                            catch (OperationCanceledException)
-                            {
-                                throw;
-                            }
-                            catch (Exception ex) when (ex is not OutOfMemoryException)
-                            {
-                                await _dispatcher.InvokeAsync(() => entry.Monitor = false);
-                                await ReportCustomOperationFailureAsync("read", entry, ex, true);
-                            }
-                        }
-
-                        var writePeriod = entry.PeriodMs <= 0 ? DefaultCustomPeriodMs : entry.PeriodMs;
-                        if (entry.Continuous && (now - entry.LastWriteUtc).TotalMilliseconds >= writePeriod)
-                        {
-                            try
-                            {
-                                var success = await WriteCustomValueSerializedAsync(entry, token);
-                                if (!success)
-                                {
-                                    throw new InvalidOperationException("The custom value is invalid for the selected type or area.");
-                                }
-
-                                var writeAt = DateTime.UtcNow;
-                                await _dispatcher.InvokeAsync(() => entry.LastWriteUtc = writeAt);
-                            }
-                            catch (OperationCanceledException)
-                            {
-                                throw;
-                            }
-                            catch (Exception ex) when (ex is not OutOfMemoryException)
-                            {
-                                await _dispatcher.InvokeAsync(() => entry.Continuous = false);
-                                await ReportCustomOperationFailureAsync("write", entry, ex, true);
-                            }
-                        }
-                    }
-
-                    await Task.Delay(PollLoopIntervalMs, token);
-                }
-            }
-            catch (OperationCanceledException)
-            {
-                _logger.LogDebug("Custom watch loop canceled");
-            }
-            catch (Exception ex) when (ex is not OutOfMemoryException)
-            {
-                _logger.LogError(ex, "Custom watch loop failed");
-                await _dispatcher.InvokeAsync(() => StatusMessage = $"Custom watch error: {ex.Message}");
-            }
-            finally
-            {
-                var restart = false;
-                lock (_customWatchLifecycleLock)
-                {
-                    if (ReferenceEquals(_customWatchCts, loopCts))
-                    {
-                        _customWatchCts = null;
-                        restart = !_disposed && ActiveProfile is { IsConnected: true } && IsCustomWatchMonitoring;
-                    }
-                }
-
-                loopCts.Dispose();
-                if (restart)
-                {
-                    StartCustomWatchMonitoring();
-                }
-            }
-        }
-
-        private static string GenerateNextName(string previousName)
-        {
-            if (string.IsNullOrWhiteSpace(previousName))
-                return "Tag0";
-
-            int i = previousName.Length - 1;
-            while (i >= 0 && char.IsDigit(previousName[i]))
-                i--;
-
-            if (i < previousName.Length - 1 &&
-                int.TryParse(previousName[(i + 1)..], NumberStyles.Integer, CultureInfo.InvariantCulture, out var num))
-            {
-                return previousName.Substring(0, i + 1) + (num + 1).ToString(CultureInfo.InvariantCulture);
-            }
-
-            return previousName + "1";
-        }
-
-        private static bool TryParseBool(string? text, out bool result)
-        {
-            if (string.IsNullOrWhiteSpace(text))
-            {
-                result = false;
-                return false;
-            }
-
-            var trimmed = text.Trim();
-
-            if (bool.TryParse(trimmed, out result))
-                return true;
-
-            if (int.TryParse(trimmed, out var value))
-            {
-                result = value != 0;
-                return true;
-            }
-
-            if (trimmed.Equals("on", StringComparison.OrdinalIgnoreCase) ||
-                trimmed.Equals("yes", StringComparison.OrdinalIgnoreCase) ||
-                trimmed.Equals("true", StringComparison.OrdinalIgnoreCase))
-            {
-                result = true;
-                return true;
-            }
-
-            if (trimmed.Equals("off", StringComparison.OrdinalIgnoreCase) ||
-                trimmed.Equals("no", StringComparison.OrdinalIgnoreCase) ||
-                trimmed.Equals("false", StringComparison.OrdinalIgnoreCase))
-            {
-                result = false;
-                return true;
-            }
-
-            result = false;
-            return false;
-        }
-
-        private static bool TryParseTrendValue(string? text, out double result)
-        {
-            if (string.IsNullOrWhiteSpace(text))
-            {
-                result = 0;
-                return false;
-            }
-
-            var trimmed = text.Trim();
-
-            if (double.TryParse(trimmed, NumberStyles.Float, CultureInfo.InvariantCulture, out result))
-                return true;
-
-            if (bool.TryParse(trimmed, out var b))
-            {
-                result = b ? 1 : 0;
-                return true;
-            }
-
-            return false;
-        }
 
         private IEnumerable<MqttTagUpdate> BuildMqttSnapshot()
         {
@@ -3563,797 +822,7 @@ namespace ModbusForge.Avalonia.ViewModels
             return updates;
         }
 
-        #endregion
 
-        #region Project Save/Load
-
-        private async Task ExportUnitIdsAsync()
-        {
-            if (_fileDialogService == null) return;
-
-            try
-            {
-                var path = await _fileDialogService.ShowSaveFileDialogAsync(
-                    "Export Unit ID Configurations",
-                    "ModbusForge Unit IDs (*.mfp;*.mui)|*.mfp;*.mui|JSON files (*.json)|*.json|All files (*.*)|*.*",
-                    "unit-id-configurations.mfp");
-
-                if (path == null) return;
-
-                var snapshot = BuildWorkspaceSnapshot();
-                var project = CreateProjectConfiguration(snapshot, "Exported Unit ID Configurations");
-                project.VisualNodes = new List<VisualNode>();
-                project.VisualConnections = new List<NodeConnection>();
-                await _fileSystem.WriteAllTextAsync(path, JsonSerializer.Serialize(project, PersistenceJsonOptions));
-                StatusMessage = $"Exported {snapshot.UnitConfigurations.Count} Unit ID configuration(s) to {Path.GetFileName(path)}.";
-            }
-            catch (Exception ex) when (ex is not (OutOfMemoryException or OperationCanceledException))
-            {
-                _logger.LogError(ex, "Error exporting Unit ID configurations");
-                StatusMessage = $"Unit ID export error: {ex.Message}";
-            }
-        }
-
-        private async Task ExportUnitIdAsync()
-        {
-            if (_fileDialogService == null || !IsServerMode) return;
-
-            try
-            {
-                var selectedUnitId = SelectedUnitId;
-                var path = await _fileDialogService.ShowSaveFileDialogAsync(
-                    $"Export Unit ID {selectedUnitId}",
-                    "ModbusForge Unit ID (*.mui)|*.mui|ModbusForge Project (*.mfp)|*.mfp|All files (*.*)|*.*",
-                    $"unit-id-{selectedUnitId}.mui");
-
-                if (path == null) return;
-
-                var snapshot = BuildWorkspaceSnapshot();
-                var project = CreateProjectConfiguration(snapshot, $"Unit ID {selectedUnitId}");
-                project.UnitConfigurations = new Dictionary<byte, UnitIdConfiguration>
-                {
-                    [selectedUnitId] = snapshot.UnitConfigurations.TryGetValue(selectedUnitId, out var configuration)
-                        ? configuration.Clone()
-                        : new UnitIdConfiguration(selectedUnitId)
-                };
-                project.VisualNodes = new List<VisualNode>();
-                project.VisualConnections = new List<NodeConnection>();
-                await _fileSystem.WriteAllTextAsync(path, JsonSerializer.Serialize(project, PersistenceJsonOptions));
-                StatusMessage = $"Unit ID {selectedUnitId} exported to {Path.GetFileName(path)}.";
-            }
-            catch (Exception ex) when (ex is not (OutOfMemoryException or OperationCanceledException))
-            {
-                _logger.LogError(ex, "Error exporting Unit ID {UnitId}", SelectedUnitId);
-                StatusMessage = $"Unit ID export error: {ex.Message}";
-            }
-        }
-
-        private async Task ImportUnitIdsAsync()
-        {
-            if (_fileDialogService == null) return;
-
-            try
-            {
-                var path = await _fileDialogService.ShowOpenFileDialogAsync(
-                    "Import Unit ID Configurations",
-                    "ModbusForge Unit IDs (*.mfp;*.mui)|*.mfp;*.mui|JSON files (*.json)|*.json|All files (*.*)|*.*");
-
-                if (path == null) return;
-
-                var json = await _fileSystem.ReadAllTextAsync(path);
-                if (TryDeserializeUnitConfigurations(json, out var configurations))
-                {
-                    var imported = 0;
-                    foreach (var pair in configurations)
-                    {
-                        if (pair.Key is < 1 or > 247 || _unitConfigurationStore.TryGetConfiguration(pair.Key, out _))
-                        {
-                            continue;
-                        }
-
-                        _unitConfigurationStore.SetConfiguration(pair.Key, pair.Value);
-                        imported++;
-                    }
-
-                    var availableIds = _unitConfigurationStore.UnitConfigurations.Keys
-                        .Where(id => id is >= 1 and <= 247)
-                        .OrderBy(id => id)
-                        .ToList();
-                    _unitConfigurationStore.PopulateAvailableUnitIds(availableIds);
-                    if (ActiveProfile != null && IsServerMode)
-                    {
-                        ActiveProfile.ServerUnitIds = string.Join(",", availableIds);
-                    }
-
-                    StatusMessage = $"Imported {imported} new Unit ID configuration(s) from {Path.GetFileName(path)}.";
-                    return;
-                }
-
-                // Backward compatibility with the original [1, 2, 5] JSON format.
-                var ids = TryDeserializeUnitIdList(json);
-                if (ids.Count == 0)
-                {
-                    StatusMessage = "No valid Unit ID configurations were found in the selected file.";
-                    return;
-                }
-
-                foreach (var id in ids)
-                {
-                    _unitConfigurationStore.GetOrCreateConfiguration(id);
-                }
-
-                _unitConfigurationStore.PopulateAvailableUnitIds(ids);
-                if (ActiveProfile != null && IsServerMode)
-                {
-                    ActiveProfile.ServerUnitIds = string.Join(",", ids);
-                }
-
-                SelectedUnitId = ids[0];
-                OnPropertyChanged(nameof(ServerUnitIds));
-                StatusMessage = $"Imported {ids.Count} Unit ID(s) from {Path.GetFileName(path)}.";
-            }
-            catch (Exception ex) when (ex is not (OutOfMemoryException or OperationCanceledException))
-            {
-                _logger.LogError(ex, "Error importing Unit ID configurations");
-                StatusMessage = $"Unit ID import error: {ex.Message}";
-            }
-        }
-
-        private async Task ImportUnitIdAsAsync()
-        {
-            if (_fileDialogService == null || _inputDialogService == null || !IsServerMode)
-            {
-                return;
-            }
-
-            try
-            {
-                var path = await _fileDialogService.ShowOpenFileDialogAsync(
-                    "Import Unit ID Configuration",
-                    "ModbusForge Unit ID (*.mui)|*.mui|ModbusForge Project (*.mfp)|*.mfp|JSON files (*.json)|*.json|All files (*.*)|*.*");
-                if (path == null) return;
-
-                var json = await _fileSystem.ReadAllTextAsync(path);
-                if (!TryDeserializeUnitConfigurations(json, out var configurations) || configurations.Count == 0)
-                {
-                    StatusMessage = "No Unit ID configurations were found in the selected file.";
-                    return;
-                }
-
-                var source = configurations.First();
-                if (!_inputDialogService.TryGetInput(
-                        "Import Unit ID As",
-                        $"Enter target Unit ID (1-247) to import Unit ID {source.Key} as:",
-                        source.Key.ToString(CultureInfo.InvariantCulture),
-                        out var input))
-                {
-                    return;
-                }
-
-                if (!byte.TryParse(input, NumberStyles.Integer, CultureInfo.InvariantCulture, out var target)
-                    || target is < 1 or > 247)
-                {
-                    StatusMessage = "Invalid target Unit ID. Enter a value between 1 and 247.";
-                    return;
-                }
-
-                var imported = source.Value.Clone();
-                imported.UnitId = target;
-                _unitConfigurationStore.SetConfiguration(target, imported);
-                var ids = _unitConfigurationStore.UnitConfigurations.Keys
-                    .Where(id => id is >= 1 and <= 247)
-                    .OrderBy(id => id)
-                    .ToList();
-                _unitConfigurationStore.PopulateAvailableUnitIds(ids);
-                if (ActiveProfile != null)
-                {
-                    ActiveProfile.ServerUnitIds = string.Join(",", ids);
-                }
-
-                SelectedUnitId = target;
-                StatusMessage = $"Imported Unit ID {source.Key} as Unit ID {target}.";
-            }
-            catch (Exception ex) when (ex is not (OutOfMemoryException or OperationCanceledException))
-            {
-                _logger.LogError(ex, "Error importing Unit ID configuration");
-                StatusMessage = $"Unit ID import error: {ex.Message}";
-            }
-        }
-
-        private async Task SaveProjectAsync()
-        {
-            if (_fileDialogService == null) return;
-
-            IsBusy = true;
-            try
-            {
-                var path = await _fileDialogService.ShowSaveFileDialogAsync(
-                    "Save ModbusForge Project",
-                    "ModbusForge Project (*.mfp)|*.mfp|JSON files (*.json)|*.json|All files (*.*)|*.*",
-                    "project.mfp");
-
-                if (path == null) return;
-
-                var snapshot = BuildWorkspaceSnapshot();
-                var project = CreateProjectConfiguration(snapshot, Path.GetFileNameWithoutExtension(path));
-                project.Profiles = _connectionManager.Profiles.ToList();
-                project.ActiveProfileId = ActiveProfile?.Id;
-                project.SelectedUnitId = snapshot.SelectedUnitId;
-                await _fileSystem.WriteAllTextAsync(path, JsonSerializer.Serialize(project, PersistenceJsonOptions));
-
-                StatusMessage = $"Saved project to {Path.GetFileName(path)}.";
-            }
-            catch (Exception ex) when (ex is not (OutOfMemoryException or OperationCanceledException))
-            {
-                _logger.LogError(ex, "Error saving project");
-                StatusMessage = $"Save error: {ex.Message}";
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        private async Task LoadProjectAsync()
-        {
-            if (_fileDialogService == null) return;
-
-            IsBusy = true;
-            try
-            {
-                var path = await _fileDialogService.ShowOpenFileDialogAsync(
-                    "Load ModbusForge Project",
-                    "ModbusForge Project (*.mfp)|*.mfp|JSON files (*.json)|*.json|All files (*.*)|*.*");
-
-                if (path == null) return;
-
-                var json = await _fileSystem.ReadAllTextAsync(path);
-                if (!TryDeserializeProject(json, out var snapshot, out var profiles, out var activeProfileId))
-                {
-                    StatusMessage = "The selected project file is empty or not a supported ModbusForge project.";
-                    return;
-                }
-
-                await _dispatcher.InvokeAsync(() =>
-                {
-                    ApplyLoadedProfiles(profiles, activeProfileId);
-                    ApplyWorkspaceSnapshot(snapshot);
-                    StatusMessage = $"Loaded project from {Path.GetFileName(path)}.";
-                });
-            }
-            catch (Exception ex) when (ex is not (OutOfMemoryException or OperationCanceledException))
-            {
-                _logger.LogError(ex, "Error loading project");
-                StatusMessage = $"Load error: {ex.Message}";
-            }
-            finally
-            {
-                IsBusy = false;
-            }
-        }
-
-        private ProjectWorkspaceSnapshot BuildWorkspaceSnapshot()
-        {
-            SyncCurrentUnitConfiguration();
-
-            var snapshot = new ProjectWorkspaceSnapshot
-            {
-                Mode = Mode,
-                ServerAddress = ActiveProfile?.IpAddress ?? "127.0.0.1",
-                Port = ActiveProfile?.Port ?? 502,
-                ServerUnitId = ServerUnitIds,
-                ClientUnitId = (byte)Math.Clamp(UnitId, 1, 247),
-                SelectedUnitId = SelectedUnitId,
-                IsServerMode = IsServerMode,
-                VisibleTabs = GetVisibleTabs(),
-                VisualNodes = VisualNodeEditorViewModel?.Nodes.ToList() ?? new List<VisualNode>(),
-                VisualConnections = VisualNodeEditorViewModel?.Connections.ToList() ?? new List<NodeConnection>()
-            };
-
-            foreach (var pair in _unitConfigurationStore.UnitConfigurations)
-            {
-                if (pair.Key is >= 1 and <= 247 && pair.Value != null)
-                {
-                    snapshot.UnitConfigurations[pair.Key] = pair.Value.Clone();
-                }
-            }
-
-            if (snapshot.UnitConfigurations.Count == 0)
-            {
-                snapshot.UnitConfigurations[snapshot.IsServerMode ? snapshot.SelectedUnitId : snapshot.ClientUnitId]
-                    = new UnitIdConfiguration(snapshot.IsServerMode ? snapshot.SelectedUnitId : snapshot.ClientUnitId);
-            }
-
-            return snapshot;
-        }
-
-        private void ApplyWorkspaceSnapshot(ProjectWorkspaceSnapshot snapshot)
-        {
-            if (snapshot == null)
-            {
-                return;
-            }
-
-            _isApplyingUnitConfiguration = true;
-            try
-            {
-                if (ActiveProfile == null)
-                {
-                    var profile = new ConnectionProfile("Default", snapshot.ServerAddress, snapshot.Port, snapshot.ClientUnitId)
-                    {
-                        Mode = string.IsNullOrWhiteSpace(snapshot.Mode) ? "Client" : snapshot.Mode,
-                        ServerUnitIds = snapshot.ServerUnitId
-                    };
-                    _connectionManager.AddProfile(profile);
-                    _connectionManager.SetActiveProfile(profile);
-                }
-                else
-                {
-                    ActiveProfile.Mode = string.IsNullOrWhiteSpace(snapshot.Mode) ? ActiveProfile.Mode : snapshot.Mode;
-                    ActiveProfile.IpAddress = string.IsNullOrWhiteSpace(snapshot.ServerAddress)
-                        ? ActiveProfile.IpAddress
-                        : snapshot.ServerAddress;
-                    if (snapshot.Port > 0)
-                    {
-                        ActiveProfile.Port = snapshot.Port;
-                    }
-
-                    ActiveProfile.ServerUnitIds = string.IsNullOrWhiteSpace(snapshot.ServerUnitId)
-                        ? ActiveProfile.ServerUnitIds
-                        : snapshot.ServerUnitId;
-                    ActiveProfile.UnitId = snapshot.ClientUnitId is >= 1 and <= 247
-                        ? snapshot.ClientUnitId
-                        : ActiveProfile.UnitId;
-                }
-
-                _unitConfigurationStore.Clear();
-                var configurations = snapshot.UnitConfigurations ?? new Dictionary<byte, UnitIdConfiguration>();
-                foreach (var pair in configurations)
-                {
-                    if (pair.Key is >= 1 and <= 247 && pair.Value != null)
-                    {
-                        var configuration = pair.Value.Clone();
-                        configuration.UnitId = pair.Key;
-                        _unitConfigurationStore.SetConfiguration(pair.Key, configuration);
-                    }
-                }
-
-                var ids = _unitConfigurationStore.UnitConfigurations.Keys
-                    .Where(id => id is >= 1 and <= 247)
-                    .OrderBy(id => id)
-                    .ToList();
-                var requested = snapshot.SelectedUnitId is >= 1 and <= 247
-                    ? snapshot.SelectedUnitId
-                    : snapshot.ClientUnitId;
-                if (ids.Count == 0)
-                {
-                    requested = requested is >= 1 and <= 247 ? requested : (byte)1;
-                    _unitConfigurationStore.GetOrCreateConfiguration(requested);
-                    ids.Add(requested);
-                }
-
-                _unitConfigurationStore.PopulateAvailableUnitIds(ids);
-                _unitConfigurationStore.SelectedUnitId = ids.Contains(requested) ? requested : ids[0];
-                ApplyCurrentUnitConfiguration();
-
-                SetVisibleTabs(snapshot.VisibleTabs);
-
-                if (VisualNodeEditorViewModel != null)
-                {
-                    VisualNodeEditorViewModel.Nodes.Clear();
-                    VisualNodeEditorViewModel.Connections.Clear();
-                    foreach (var node in snapshot.VisualNodes ?? new List<VisualNode>())
-                    {
-                        VisualNodeEditorViewModel.Nodes.Add(node);
-                    }
-
-                    foreach (var connection in snapshot.VisualConnections ?? new List<NodeConnection>())
-                    {
-                        VisualNodeEditorViewModel.Connections.Add(connection);
-                    }
-                }
-            }
-            finally
-            {
-                _isApplyingUnitConfiguration = false;
-            }
-
-            OnPropertyChanged(nameof(ActiveProfile));
-            OnPropertyChanged(nameof(UnitId));
-            OnPropertyChanged(nameof(Mode));
-            OnPropertyChanged(nameof(ServerUnitIds));
-            OnPropertyChanged(nameof(EffectiveUnitId));
-            OnPropertyChanged(nameof(CurrentConfig));
-            OnPropertyChanged(nameof(CustomEntries));
-            OnPropertyChanged(nameof(AvailableUnitIds));
-            ReadAllCustomNowCommand.NotifyCanExecuteChanged();
-            ExportUnitIdCommand.NotifyCanExecuteChanged();
-            ImportUnitIdAsCommand.NotifyCanExecuteChanged();
-        }
-
-        private void ApplyLoadedProfiles(
-            IReadOnlyList<ConnectionProfile>? profiles,
-            string? activeProfileId)
-        {
-            if (profiles == null || profiles.Count == 0)
-            {
-                return;
-            }
-
-            // Loading a project replaces the whole profile set - stop any pending reconnects.
-            // (The per-profile "still in the manager" check in HandleProfileDisconnected
-            // prevents auto-reconnect for the old profiles.)
-            StopAutoReconnect();
-            foreach (var connectedProfile in _connectionManager.Profiles.Where(profile => profile.IsConnected).ToList())
-            {
-                _ = _connectionManager.DisconnectProfileAsync(connectedProfile);
-            }
-
-            _connectionManager.Profiles.Clear();
-            foreach (var source in profiles)
-            {
-                source.IsConnected = false;
-                source.IsActive = false;
-                source.Status = "Disconnected";
-                _connectionManager.AddProfile(source);
-            }
-
-            var active = _connectionManager.Profiles.FirstOrDefault(profile => profile.Id == activeProfileId)
-                         ?? _connectionManager.Profiles.FirstOrDefault();
-            if (active != null)
-            {
-                _connectionManager.SetActiveProfile(active);
-            }
-
-            _connectionManager.SaveProfiles();
-        }
-
-        private static AvaloniaProjectConfiguration CreateProjectConfiguration(
-            ProjectWorkspaceSnapshot snapshot,
-            string projectName)
-        {
-            var project = new AvaloniaProjectConfiguration
-            {
-                ProjectInfo = new ProjectInfo
-                {
-                    Name = string.IsNullOrWhiteSpace(projectName) ? "ModbusForge Project" : projectName,
-                    Version = Assembly.GetExecutingAssembly().GetName().Version?.ToString() ?? string.Empty,
-                    Modified = DateTime.Now
-                },
-                GlobalSettings = new GlobalSettings
-                {
-                    Mode = snapshot.Mode,
-                    ServerAddress = snapshot.ServerAddress,
-                    Port = snapshot.Port,
-                    ServerUnitId = snapshot.ServerUnitId,
-                    ClientUnitId = snapshot.ClientUnitId,
-                    VisibleTabs = snapshot.VisibleTabs?.ToList() ?? new List<string>()
-                },
-                UnitConfigurations = snapshot.UnitConfigurations
-                    .ToDictionary(pair => pair.Key, pair => pair.Value.Clone()),
-                VisualNodes = snapshot.VisualNodes?.ToList() ?? new List<VisualNode>(),
-                VisualConnections = snapshot.VisualConnections?.ToList() ?? new List<NodeConnection>()
-            };
-
-            return project;
-        }
-
-        private static bool TryDeserializeProject(
-            string json,
-            out ProjectWorkspaceSnapshot snapshot,
-            out IReadOnlyList<ConnectionProfile>? profiles,
-            out string? activeProfileId)
-        {
-            snapshot = new ProjectWorkspaceSnapshot();
-            profiles = null;
-            activeProfileId = null;
-
-            if (string.IsNullOrWhiteSpace(json))
-            {
-                return false;
-            }
-
-            try
-            {
-                using var document = JsonDocument.Parse(json);
-                var root = document.RootElement;
-
-                if (TryGetProperty(root, "globalSettings", out _))
-                {
-                    var project = JsonSerializer.Deserialize<AvaloniaProjectConfiguration>(json, PersistenceJsonOptions);
-                    if (project == null || project.UnitConfigurations == null)
-                    {
-                        return false;
-                    }
-
-                    profiles = project.Profiles;
-                    activeProfileId = project.ActiveProfileId;
-                    snapshot = SnapshotFromProjectConfiguration(project, project.SelectedUnitId, null);
-                    return snapshot.UnitConfigurations.Count > 0 || snapshot.VisualNodes.Count > 0 || profiles.Count > 0;
-                }
-
-                if (TryGetProperty(root, "unitConfigurations", out _)
-                    && TryGetProperty(root, "mode", out _))
-                {
-                    snapshot = JsonSerializer.Deserialize<ProjectWorkspaceSnapshot>(json, PersistenceJsonOptions)
-                               ?? new ProjectWorkspaceSnapshot();
-                    return snapshot.UnitConfigurations.Count > 0;
-                }
-
-                // Prior Avalonia builds wrote AppConfiguration directly. Keep those
-                // files loadable while all new files use ProjectConfiguration plus
-                // the profile extension below.
-                var legacy = JsonSerializer.Deserialize<AppConfiguration>(json, PersistenceJsonOptions);
-                if (legacy == null)
-                {
-                    return false;
-                }
-
-                profiles = legacy.Profiles;
-                activeProfileId = legacy.ActiveProfileId;
-                var active = legacy.Profiles?.FirstOrDefault(profile => profile.Id == legacy.ActiveProfileId)
-                             ?? legacy.Profiles?.FirstOrDefault();
-                var clientId = legacy.UnitId is >= 1 and <= 247
-                    ? legacy.UnitId
-                    : (active?.UnitId is >= 1 and <= 247 ? active.UnitId : (byte)1);
-                var configuration = new UnitIdConfiguration(clientId);
-                foreach (var entry in legacy.CustomEntries ?? new List<CustomEntry>())
-                {
-                    configuration.CustomEntries.Add(entry);
-                }
-
-                configuration.RegisterSettings.RegisterStart = legacy.StartAddress;
-                configuration.RegisterSettings.RegisterCount = legacy.RegisterCount;
-                configuration.RegisterSettings.RegistersGlobalType = legacy.GlobalType ?? "int";
-                configuration.RegisterSettings.RegistersSwapBytes = legacy.SwapBytes;
-                configuration.RegisterSettings.RegistersSwapWords = legacy.SwapWords;
-                snapshot = new ProjectWorkspaceSnapshot
-                {
-                    Mode = legacy.Mode ?? active?.Mode ?? "Client",
-                    ServerAddress = legacy.ServerAddress ?? active?.IpAddress ?? "127.0.0.1",
-                    Port = legacy.Port > 0 ? legacy.Port : active?.Port ?? 502,
-                    ServerUnitId = active?.ServerUnitIds ?? "1",
-                    ClientUnitId = clientId,
-                    SelectedUnitId = clientId,
-                    IsServerMode = string.Equals(legacy.Mode ?? active?.Mode, "Server", StringComparison.OrdinalIgnoreCase),
-                    UnitConfigurations = new Dictionary<byte, UnitIdConfiguration> { [clientId] = configuration },
-                    VisualNodes = legacy.VisualNodes ?? new List<VisualNode>(),
-                    VisualConnections = legacy.VisualConnections ?? new List<NodeConnection>()
-                };
-                return true;
-            }
-            catch (JsonException)
-            {
-                return false;
-            }
-        }
-
-        private static ProjectWorkspaceSnapshot SnapshotFromProjectConfiguration(
-            ProjectConfiguration project,
-            byte? selectedUnitId,
-            IReadOnlyList<string>? visibleTabsOverride)
-        {
-            var global = project.GlobalSettings ?? new GlobalSettings();
-            var configurations = new Dictionary<byte, UnitIdConfiguration>();
-            foreach (var pair in project.UnitConfigurations ?? new Dictionary<byte, UnitIdConfiguration>())
-            {
-                if (pair.Key is >= 1 and <= 247 && pair.Value != null)
-                {
-                    var configuration = pair.Value.Clone();
-                    configuration.UnitId = pair.Key;
-                    configurations[pair.Key] = configuration;
-                }
-            }
-
-            var clientId = global.ClientUnitId is >= 1 and <= 247 ? global.ClientUnitId : (byte)1;
-            var selected = selectedUnitId is >= 1 and <= 247
-                ? selectedUnitId.Value
-                : (configurations.ContainsKey(clientId) ? clientId : configurations.Keys.FirstOrDefault((byte)1));
-            return new ProjectWorkspaceSnapshot
-            {
-                Mode = global.Mode ?? "Client",
-                ServerAddress = global.ServerAddress ?? "127.0.0.1",
-                Port = global.Port > 0 ? global.Port : 502,
-                ServerUnitId = global.ServerUnitId ?? "1",
-                ClientUnitId = clientId,
-                SelectedUnitId = selected,
-                IsServerMode = string.Equals(global.Mode, "Server", StringComparison.OrdinalIgnoreCase),
-                UnitConfigurations = configurations,
-                VisibleTabs = visibleTabsOverride?.ToList() ?? global.VisibleTabs?.ToList() ?? new List<string>(),
-                VisualNodes = project.VisualNodes ?? new List<VisualNode>(),
-                VisualConnections = project.VisualConnections ?? new List<NodeConnection>()
-            };
-        }
-
-        private static bool TryDeserializeUnitConfigurations(
-            string json,
-            out Dictionary<byte, UnitIdConfiguration> configurations)
-        {
-            configurations = new Dictionary<byte, UnitIdConfiguration>();
-            if (string.IsNullOrWhiteSpace(json))
-            {
-                return false;
-            }
-
-            try
-            {
-                using var document = JsonDocument.Parse(json);
-                var root = document.RootElement;
-                if (!TryGetProperty(root, "unitConfigurations", out _))
-                {
-                    return false;
-                }
-
-                var project = JsonSerializer.Deserialize<ProjectConfiguration>(json, PersistenceJsonOptions);
-                if (project?.UnitConfigurations == null)
-                {
-                    return false;
-                }
-
-                foreach (var pair in project.UnitConfigurations)
-                {
-                    if (pair.Key is >= 1 and <= 247 && pair.Value != null)
-                    {
-                        var configuration = pair.Value.Clone();
-                        configuration.UnitId = pair.Key;
-                        configurations[pair.Key] = configuration;
-                    }
-                }
-
-                return configurations.Count > 0;
-            }
-            catch (JsonException)
-            {
-                return false;
-            }
-        }
-
-        private static List<byte> TryDeserializeUnitIdList(string json)
-        {
-            try
-            {
-                return (JsonSerializer.Deserialize<List<byte>>(json, PersistenceJsonOptions) ?? new List<byte>())
-                    .Where(id => id is >= 1 and <= 247)
-                    .Distinct()
-                    .OrderBy(id => id)
-                    .ToList();
-            }
-            catch (JsonException)
-            {
-                return new List<byte>();
-            }
-        }
-
-        private static bool TryGetProperty(JsonElement element, string name, out JsonElement value)
-        {
-            if (element.ValueKind != JsonValueKind.Object)
-            {
-                value = default;
-                return false;
-            }
-
-            foreach (var property in element.EnumerateObject())
-            {
-                if (string.Equals(property.Name, name, StringComparison.OrdinalIgnoreCase))
-                {
-                    value = property.Value;
-                    return true;
-                }
-            }
-
-            value = default;
-            return false;
-        }
-
-        private static readonly JsonSerializerOptions PersistenceJsonOptions = new()
-        {
-            WriteIndented = true,
-            PropertyNameCaseInsensitive = true,
-            PropertyNamingPolicy = JsonNamingPolicy.CamelCase
-        };
-
-        private sealed class AvaloniaProjectConfiguration : ProjectConfiguration
-        {
-            public List<ConnectionProfile> Profiles { get; set; } = new();
-            public string? ActiveProfileId { get; set; }
-            public byte SelectedUnitId { get; set; } = 1;
-        }
-
-        private void ToggleTheme()
-        {
-            _themeService?.ToggleTheme();
-        }
-
-        private async Task CheckForUpdatesAsync()
-        {
-            if (_updateService == null) return;
-
-            try
-            {
-                var result = await _updateService.CheckForUpdateAsync();
-                if (result.IsUpdateAvailable)
-                {
-                    var msg = $"A newer version is available: {result.LatestVersion}\nCurrent: {result.CurrentVersion}\n\nDownload and install it now?";
-                    if (_messageBoxService != null)
-                    {
-                        var dialogResult = await _messageBoxService.ShowAsync(msg, "Update Available", DialogButton.YesNo, DialogIcon.Question);
-                        if (dialogResult == DialogResult.Yes)
-                        {
-                            if (string.IsNullOrWhiteSpace(result.AssetDownloadUrl))
-                            {
-                                OpenUrl(result.ReleaseUrl);
-                                return;
-                            }
-
-                            var installerPath = Path.Combine(
-                                Path.GetTempPath(),
-                                $"ModbusForge-{result.LatestVersion}-setup.exe");
-                            var progress = new Progress<double>(value =>
-                            {
-                                StatusMessage = $"Downloading update... {value:P0}";
-                            });
-
-                            StatusMessage = "Downloading update...";
-                            var downloaded = await _updateService.DownloadInstallerAsync(
-                                result.AssetDownloadUrl,
-                                installerPath,
-                                progress);
-                            if (!downloaded)
-                            {
-                                StatusMessage = "Update download failed.";
-                                return;
-                            }
-
-                            StatusMessage = "Launching update installer...";
-                            _updateService.LaunchInstaller(installerPath);
-                            _applicationLifetime?.Shutdown();
-                        }
-                    }
-                    else
-                    {
-                        StatusMessage = $"Update available: {result.LatestVersion}";
-                    }
-                }
-                else
-                {
-                    StatusMessage = $"Up to date ({result.CurrentVersion}).";
-                }
-            }
-            catch (Exception ex)
-            {
-                StatusMessage = $"Update check failed: {ex.Message}";
-                _logger.LogWarning(ex, "Update check failed");
-            }
-        }
-
-        private void OpenUrl(string url)
-        {
-            // Only open absolute http(s) URLs - never an arbitrary string (which
-            // UseShellExecute would hand to the shell as a document/protocol).
-            if (!Uri.TryCreate(url, UriKind.Absolute, out var uri) ||
-                (uri.Scheme != Uri.UriSchemeHttp && uri.Scheme != Uri.UriSchemeHttps))
-            {
-                _logger.LogWarning("Refusing to open non-http(s) URL: {Url}", url);
-                StatusMessage = "The update URL is invalid.";
-                return;
-            }
-
-            try
-            {
-                System.Diagnostics.Process.Start(new System.Diagnostics.ProcessStartInfo
-                {
-                    FileName = url,
-                    UseShellExecute = true
-                });
-            }
-            catch (Exception ex)
-            {
-                _logger.LogWarning(ex, "Failed to open URL {Url}", url);
-                StatusMessage = $"Could not open the browser: {ex.Message}";
-            }
-        }
-
-        #endregion
 
         public void Dispose()
         {
@@ -4396,5 +865,6 @@ namespace ModbusForge.Avalonia.ViewModels
             // Release the per-entry PropertyChanged hooks so the custom entries can be collected.
             UnhookCustomEntries();
         }
+
     }
 }
